% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusterAnnotations.R
\name{addClusterAnnotations}
\alias{addClusterAnnotations}
\title{Add Cluster Annotations to Seurat Object}
\usage{
addClusterAnnotations(se, AnnoDataframe, BarcodeName, AnnoName)
}
\arguments{
\item{se}{A Seurat object.}

\item{AnnoDataframe}{A dataframe containing cell or spatial barcodes and their corresponding annotations.}

\item{BarcodeName}{Name of the column in \code{AnnoDataframe} containing barcodes (must match \code{colnames(se)}).}

\item{AnnoName}{Name of the column in \code{AnnoDataframe} containing cluster annotations.}
}
\value{
A Seurat object with added metadata column \code{seurat_clusters}.
}
\description{
This function extracts barcode and annotation columns from a provided dataframe
and adds them directly to the metadata of a Seurat object under the column \code{seurat_clusters}.
It ensures that annotations are correctly aligned by barcode.
}
\examples{
seurat_clusters <- read.csv("data/annotations/seurat_clusters.csv", sep = ";")
se <- addClusterAnnotations(se, AnnoDataframe = seurat_clusters, BarcodeName = "Barcode", AnnoName = "seurat_clusters")
}
