% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesfunctions.R
\name{addyear}
\alias{addyear}
\title{Function to analyse date column in the dataset and add a column containing only the year}
\usage{
addyear(games)
}
\arguments{
\item{games}{The "games" dataset}
}
\value{
games The "games" dataset with an added column "year"
}
\description{
Function to analyse date column in the dataset and add a column containing only the year
}
