% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{appdetails}
\alias{appdetails}
\alias{stf_appdetails}
\alias{cmd_appdetails}
\alias{steamspy_appdetails}
\title{Get app details}
\usage{
stf_appdetails(
  appid,
  client = NULL,
  filters = NULL,
  country_code = "US",
  language = "english"
)

cmd_appdetails(appid)

steamspy_appdetails(appid)
}
\arguments{
\item{appid}{Application ID referencing a Steam application. There are
generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query appIDs programmatically, e.g. using \code{\link{search_apps}}}
}}

\item{client}{Unknown.}

\item{filters}{Keys to include in the output. Can be any key that
is returned by this function. If \code{basic}, includes the following
default keys: \code{type}, \code{name}, \code{steam_appid},
\code{required_age}, \code{dlc}, \code{detailed_description},
\code{about_the_game}, \code{short_description}, \code{supported_languages},
\code{header_image}, \code{website}, \code{pc_requirements},
\code{mac_requirements}, \code{linux_requirements}. \code{filters} does
not accept any of these keys as input.}
}
\value{
\describe{
 \item{\code{stf_appdetails}}{A nested list containing the following keys:
 \itemize{
   \item{\code{type}: Type of application (game, dlc, demo, advertising,
   mod, or video)}
   \item{\code{name}: Name of the application}
   \item{\code{steam_appid}: appID of the application}
   \item{\code{required_age}: Required age to view the application}
   \item{\code{controller_support}: Whether the app has controller support
   (partial or full). May be omitted.}
   \item{\code{is_free}: Whether the game is free to play}
   \item{\code{dlc}: Array of appIDs of DLCs associated with this application.}
   \item{\code{detailed_description}: A detailed description in HTML}
   \item{\code{about_the_game}: "About the game" description in HTML}
   \item{\code{short_description}: Short description in HTML, usually found
   in the application preview}
   \item{\code{fullgame}: Name and appID of the full game. Sometimes
   found in demos or movies.}
   \item{\code{supported_languages}: Languages that the application supports,
   in HTML.}
   \item{\code{header_image}: URL to the header image.}
   \item{\code{capsule_image}: URL to the capsule image.}
   \item{\code{capsule_imagev5}: URL to a different capsule image.}
   \item{\code{website}: URL to the application website.}
   \item{\code{pc_requirements}: Minimum and recommended requirements
   for Windows, in HTML}
   \item{\code{mac_requirements}: Minimum and recommended requirements
   for Mac OS, in HTML}
   \item{\code{linux_requirements}: Minimum and recommended requirements
   for Linux, in HTML}
   \item{\code{developers}: Names of the developers.}
   \item{\code{publishers}: Names of the publishers.}
   \item{\code{demos}: List of appIDs which are demos of this application.}
   \item{\code{price_overview}: List containing the initial and final
   price as well as price discounts. Omitted for free to play games.}
   \item{\code{packages}: List of packageIDs that include this game.}
   \item{\code{package_groups}: Dataframe specifying purchase options.}
   \item{\code{platforms}: Specifies whether the application is compatible
   with linux and mac.}
   \item{\code{metacritic}: Metacritic score and URL.}
   \item{\code{categories}: Store categories of the application. May be
   omitted.}
   \item{\code{genres}: Genres of the application. May be omitted.}
   \item{\code{screenshots}: A dataframe containing the paths to all
   showcase screenshots. Omitted if no screenshots exist.}
   \item{\code{movies}: A dataframe containing the paths to all
   showcase movies. Omitted if no movies exist.}
   \item{\code{recommendations}: A list containing the total number
   of recommendations.}
   \item{\code{recommendations}: A list containing the total and
   highlighted achievements.}
   \item{\code{release_date}: A list containing the release date and
   whether the application is coming soon.}
   \item{\code{support_info}: A list containing support URL and email.}
   \item{\code{background}: URL to the background image.}
   \item{\code{background_raw}: URL to the raw background image.}
   \item{\code{content_descriptors}: Content descriptor codes given to
   this application. Also includes a content descriptor description.
   A list of content descriptors is included in
   \code{\link{content_descriptors}}.}
   \item{\code{ratings}: A list containing maturity ratings.}
 }
 }
 \item{\code{steamspy_appdetails}}{A dataframe containing information
 about name, developer, publisher, reviews, concurrent players, genres,
 tags and estimated owners.}

 \item{\code{cmd_appdetails}}{A nested list containing information
 listed \href{https://www.steamcmd.net/}{here}.}
}
}
\description{
Retrieve detailed information on an application in the steam store.

\itemize{
 \item{\code{stf_appdetails} uses Steam's storefront API and returns
 detailed store-related information}
 \item{\code{steamspy_appdetails} uses the SteamSpy unofficial API and
 includes third-party estimates on playtime, genres, CCU and ownership}
 \item{\code{cmd_appdetails} retrieves app details from
 \href{https://developer.valvesoftware.com/wiki/SteamCMD}{SteamCMD}, which
 is largely technical}
}
}
\note{
\code{stf_appdetails} is rate-limited at 200 requests per 5 minutes.

\code{cmd_appdetails} is not rate-limited.

\code{steamspy_appdetails} is rate-limited at 1 request per second.
}
\examples{
\donttest{# returns the detailed description about team fortress
stf_appdetails(440, filters = "detailed_description")

# returns the metacritic score and total number of reviews of Counter-Strike
stf_appdetails(10, filters = c("recommendations", "metacritic"))

# returns steamspy data
steamspy_appdetails(440)

# returns steamcmd data
cmd_appdetails(440)}
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{stf_appdetails}      \tab no \tab no\cr
    \code{steamspy_appdetails} \tab no \tab no\cr
    \code{cmd_appdetails}      \tab no \tab no
}




}
