% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_credentials}
\alias{auth_credentials}
\alias{auth_qr}
\title{Session authentication}
\usage{
auth_credentials(
  username,
  password = openssl::askpass,
  shared_secret = NULL,
  persistent = FALSE,
  friendly_name = NULL,
  details = NULL
)

auth_qr(friendly_name = NULL, device_details = NULL)
}
\arguments{
\item{username}{Account name of the user to be authenticated.}

\item{password}{A function that safely prompts the Steam password. Defaults
to \code{\link[askpass]{askpass}} which obfuscates the input. The password
is RSA encrypted before transmission to the Steam servers.}

\item{shared_secret}{A 5-character string representing the one-time Steam
Guard code needed to confirm the login. Note that the Steam Guard code
is regenerated every 30 seconds. If \code{shared_secret} is \code{NULL},
authenticates using two-factor authentication. Two-factor confirmation
requires the manual confirmation of the login request in the Steam Guard
mobile app.}

\item{persistent}{Whether to start a persistent or ephemeral session.}

\item{friendly_name}{Name of the session. Used to track authorized devices
in Steam Guard.}

\item{details}{Arbitrary details about the device attempting to
authenticate. Will be shown in Steam Guard.}
}
\value{
Information about the authenticated session.
}
\description{
Initializes an authorized Steam session. Some functions in both the
Web API and internal API require the session to be authenticated. This
concerns all API methods that access personal account information.
\code{steam_login} essentially signs in a user programmatically.

\code{show_steam_qr} authenticates by showing a QR code that can be
scanned using the Steam mobile app. The QR code is shown as an R plot
and refreshes every 5 seconds.
}
\note{
Session authentication is only possible in interactive mode because it
requires the manual insertion of Steam guard codes, confirmation of
login requests in the mobile app, or the scanning of QR codes. All of these
actions are not suitable for batch processing.
}
\examples{
\dontrun{
user <- "gabelogannewell"
auth_credentials(user)

# use a different password method
auth_credentials(user, password = rstudioapi::askForPassword)

# set a friendly name to identify this session
auth_credentials(user, friendly_name = "R session")

# sign in using a QR code
auth_qr()
}
}
