% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesfunctions.R
\name{avgpriceperyear}
\alias{avgpriceperyear}
\title{Function to visualise the average game price per year
Genres include: Action, Adventure, Casual, Indie, MassivelyMultiplayer, Racing, RPG, Simulation, Sports, Strategy}
\usage{
avgpriceperyear(games, selectedgenres = NULL)
}
\arguments{
\item{games}{The "games" dataset}

\item{selectedgenres}{Vector of genres to be visualised, OPTIONAL (all genres are selected if this parameter is not provided)}
}
\value{
avgpriceplot a ggplot object
}
\description{
Function to visualise the average game price per year
Genres include: Action, Adventure, Casual, Indie, MassivelyMultiplayer, Racing, RPG, Simulation, Sports, Strategy
}
\examples{
avgpriceperyear(games, selectedgenres = c("Action", "Indie", "Simulation"))
}
