% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_p_values.R
\name{compute_p_values}
\alias{compute_p_values}
\title{Compute P-Values for Permutation Tests}
\usage{
compute_p_values(
  structure_data,
  perm_data,
  cluster_col,
  permutation_nr = 10000,
  n_tests = 18
)
}
\arguments{
\item{structure_data}{The quantile data for the structure.}

\item{perm_data}{The permutation data frame.}

\item{cluster_col}{The column in the permutation data for the cluster.}

\item{permutation_nr}{Number of permutations (default = 10000).}

\item{n_tests}{Number of tests for adjustment (default = 18).}
}
\value{
A matrix with p-values and adjusted p-values.
}
\description{
This function computes p-values and adjusted p-values for given clusters.
}
