% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_structure_quantile.R
\name{compute_structure_quantile}
\alias{compute_structure_quantile}
\title{Compute Quantile for Structures}
\usage{
compute_structure_quantile(metadata, quantile_value = 0.99)
}
\arguments{
\item{metadata}{The metadata data frame.}

\item{quantile_value}{The quantile to compute (default = 0.99).}
}
\value{
A data frame with computed quantiles and ranks.
}
\description{
This function computes the quantiles for the given metadata structure.
}
