% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{query_events}
\alias{query_events}
\alias{get_best_events}
\alias{get_event_details}
\title{Events}
\usage{
query_events(
  min_time = NULL,
  max_time = Sys.time(),
  ascending = FALSE,
  max_results = NULL,
  populate = FALSE,
  app_types = NULL,
  event_types = NULL,
  collection_id = NULL,
  sale_id = NULL
)

get_best_events(
  language = "english",
  include_steam_blog = TRUE,
  recency = 0,
  only_game_updates = FALSE
)

get_event_details(gid, clanid)
}
\arguments{
\item{min_time, max_time}{Time range for which to include events.}

\item{ascending}{Whether to sort event dates in an ascending order.}

\item{max_results}{Maximum number of events to return. A minimum of 10
documents will always be returned. The maximum number of documents is
restricted to 500.}

\item{populate}{Number of events to populate with additional event data.
Adds a a field \code{events} to the data that essentially contains the
output of \code{get_event_details} for the specified number of events.}

\item{app_types}{Vector of app types to be returned. Can include
\code{library}, \code{wishlist}, \code{following}, \code{recommended},
\code{steam}, \code{curator} and \code{featured}.}

\item{event_types}{Vector of event types. Can include a combination of
numbers representing event types. That's all I know.}

\item{collection_id}{Unknown.}

\item{sale_id}{Unknown.}

\item{language}{Language in which event data should be returned.}

\item{include_steam_blog}{Whether events from the Steam blog should be
included.}

\item{recency}{Specifies the number of days that game must have been last
played to be included in the query.}

\item{only_game_updates}{Whether to filter out all events that are not
game updates.}

\item{gid}{List of event IDs for which to provide details.}

\item{clanid}{List of group IDs for which the events in \code{gid} were
created.}
}
\value{
\code{query_events} returns a list of dataframes. \code{documents}
contains information on the event documents. \code{apps} contains information
about the event apps. If authenticated, also shows information about last
playtime if applicable. \code{clans} shows information about the group of
each event. If \code{populate} is larger than 0, also shows an \code{events}
field, that shows resolved information about each document including their
text body.

\code{get_best_events} and \code{get_event_details} essentially return a
list in the same format as the \code{events} field in \code{query_events}.
}
\description{
Fetches data about news and events from the Steam News Hub.

\code{query_events} queries all events based on arbitrary filters.
\code{get_best_events} fetches only the best events, whatever that means.
\code{get_event_details} returns details about a number of events within
a number of groups.
}
\examples{
\dontrun{
# set a time filter
query_events(
  min_time = as.POSIXct("2023-01-01"),
  max_time = as.POSIXct("2023-01-31")
)

# get event details about first 20 events
query_events(populate = 20)

# show only events of games in library
query_events(app_types = "library")
}
}
