% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_subset.R
\name{extract_metadata_subset}
\alias{extract_metadata_subset}
\title{Extract Metadata Subset}
\usage{
extract_metadata_subset(se_meta, pattern)
}
\arguments{
\item{se_meta}{The metadata of the Seurat object.}

\item{pattern}{The pattern to search in column names.}
}
\value{
A data frame with the subset metadata.
}
\description{
This function extracts metadata for a specific feature pattern.
}
