% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterMarkers.R
\name{findClusterMarkers}
\alias{findClusterMarkers}
\title{Find cluster markers}
\usage{
findClusterMarkers(se, nfactors, dims = NULL, M = 10)
}
\arguments{
\item{se}{A Seurat object}

\item{nfactors}{Number of factors to use for NMF}

\item{dims}{Dimensions to use for clustering}

\item{M}{Number of top markers to return for each cluster}
}
\value{
A list containing marker genes for each cluster
}
\description{
Find cluster markers
}
\examples{
clusterMarkers <- findClusterMarkers(se, nfactors = 10, dims = 1:10, M = 10)
}
