% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_permutation_matrix_parallel.R
\name{generate_permutation_matrix_parallel}
\alias{generate_permutation_matrix_parallel}
\title{Generate a permutation matrix}
\usage{
generate_permutation_matrix_parallel(
  se,
  gnum = 50,
  permutation_nr = 10000,
  workers = 10,
  cluster_col = "seurat_clusters"
)
}
\arguments{
\item{se}{A Seurat object}

\item{gnum}{Number of genes to sample for each permutation}

\item{permutation_nr}{Number of permutations}

\item{workers}{Number of parallel workers}

\item{cluster_col}{The name of the column in the metadata that contains cluster information}
}
\value{
A data frame containing the permutation matrix
}
\description{
Generate a permutation matrix
}
\examples{
perm.mat <- generate_permutation_matrix_parallel(se, gnum = 50, permutation_nr = 100, workers = 4)
}
