% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_permutation_matrix.R
\name{generate_random_gene_sets}
\alias{generate_random_gene_sets}
\title{Generate random gene sets for permutation analysis}
\usage{
generate_random_gene_sets(genes, gnum, permutation_nr)
}
\arguments{
\item{genes}{A character vector containing the genes from which to sample.}

\item{gnum}{Number of genes to sample for each permutation.}

\item{permutation_nr}{Number of random gene sets to generate.}
}
\value{
A list containing one randomly sampled gene set per permutation.
}
\description{
Generates a list of randomly sampled gene sets to be used for permutation
analysis. A separate seed is used for each permutation to make the
generated gene sets reproducible.
}
