% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app_list}
\alias{get_app_list}
\alias{get_games_by_ccu}
\alias{get_most_played_games}
\alias{get_top_releases}
\title{App list}
\usage{
get_app_list()

get_games_by_ccu(
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE
)

get_most_played_games()

get_top_releases()
}
\arguments{
\item{language}{ISO639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{elanguage}{Language identifier of the returned information.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{steam_realm}{Number describing the Steam realm. A value of 1
indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.}

\item{include}{List of extra information to include. Can be one or several
of the following: \code{release}, \code{platforms},
\code{all_purchase_options}, \code{screenshots}, \code{trailers},
\code{ratings}, \code{tag_count}, \code{reviews}, \code{basic_info},
\code{supported_languages}, \code{full_description}, \code{included_items}
and \code{assets_without_overrides}. If \code{all}, returns all information.}

\item{apply_user_filters}{Unknown.}
}
\description{
Get a list of (all) applications in the Steam store.

\code{get_app_list} fetches all appIDs in the Steam store along with
their names.

\code{get_games_by_ccu} fetches a list of the most played games by
concurrent users (CCU).

\code{get_most_played_games} fetches a list of the most played games of
all time.

\code{get_top_releases} fetches a list of the most played new releases
in the Steam store.
}
\examples{
\dontrun{
get_all_list()

# get most played games
get_games_by_ccu()

# get most popular games of all time
get_most_played_games()

# get the best releases
get_top_releases()
}
}
\seealso{
\code{\link{steamspy}} for a similar approach by the SteamSpy API
}
