% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{get_asset_prices}
\alias{get_asset_prices}
\alias{get_asset_info}
\title{Assets}
\usage{
get_asset_prices(appid, currency = NULL, language = "english")

get_asset_info(
  appid,
  classids,
  class_count = 1L,
  language = "english",
  instanceids = NULL
)
}
\arguments{
\item{appid}{A single appID of an application in the Steam store.}

\item{currency}{ISO 4217 code to filter by currency. Defaults to all
available currencies.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{classids}{A vector of classIDs of the assets to retrieve.}

\item{instanceids}{A vector of instanceIDs corresponding to the}
}
\value{
\describe{
 \item{\code{get_asset_prices}}{A dataframe containg information on assert
 prices. \code{name}, \code{date}, \code{class}, and \code{classid} provide
 metadata about assets. Optionally, tags are provided in \code{tags} and
 \code{tag_ids}. \code{class}, \code{tags}, and \code{tag_ids} are nested
 as dataframes or lists as they can contain values of arbitrary length.
 \code{prices.\*} contain the current prices for
 the currencies specified in \code{currency}. If an asset is on sale,
 \code{original_prices.\*} show the original prices.}

 \item{\code{get_asset_info}}{A dataframe with each row representing
 asset info on each provided classID. \code{\*_url} columns show
 extensions to the URL \url{http://cdn.steamcommunity.com/economy/image/}.
 \code{name}, \code{market_hash_name} and \code{market_name} show various
 types of asset names. If an asset has received a fraud warning, the
 column \code{fraudwarnings} can contain their warning strings. The
 \code{descriptions} columns contain information about asset attributes,
 \code{actions} about action URLs, and \code{tags} about asset tags.}
}
}
\description{
Retrieve information about assets and their prices. An asset is an item
from the Steam economy that can be traded. For more information about
assets, see the \href{https://partner.steamgames.com/doc/store/assets}{Steamworks documentation}.
}
\examples{
\dontrun{
# show asset prices for Team Fortress and all available currencies
get_asset_prices(440)

# show the same assets, but return only EUR
get_asset_prices(440, currency = "EUR")

# a list of valid currency codes can be retrieved using
# steamkit_enum
steamkit_enum("CurrencyCode", type = "SteamLanguage")

# get_asset_prices returns classIDs for each asset. These classIDs can
# be used to retrieve detailed information on an asset.
get_asset_class(730, c("5710093913", "5189384166"))
}
}
\references{
\url{https://wiki.teamfortress.com/wiki/WebAPI/GetAssetPrices}

\url{https://wiki.teamfortress.com/wiki/WebAPI/GetAssetClassInfo}
}
\seealso{
\code{\link{steamkit_enum}} to retrieve all currency codes

\code{\link{query_loyalty_rewards}} for loyalty assets
}
