% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_top_sellers.R
\name{get_country_list}
\alias{get_country_list}
\title{Get countries}
\usage{
get_country_list(language = "english")
}
\arguments{
\item{language}{Language string of the returned information.}
}
\value{
A dataframe containing country codes and names
}
\description{
Retrieve a list of all countries of Steam and their country codes
}
\examples{
\dontrun{
get_country_list(language = "german")
}
}
