% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{get_discovery_queue}
\alias{get_discovery_queue}
\alias{get_discovery_settings}
\alias{get_discovery_skipped}
\title{Discovery queue}
\usage{
get_discovery_queue(
  queue_type = NULL,
  country_code = "US",
  rebuild_queue = FALSE,
  settings_changed = FALSE,
  os_win = FALSE,
  os_mac = FALSE,
  os_linux = FALSE,
  full_controller_support = FALSE,
  native_steam_controller = FALSE,
  include_coming_soon = FALSE,
  excluded_tagids = NULL,
  exclude_early_access = FALSE,
  exclude_videos = FALSE,
  exclude_software = FALSE,
  exclude_dlc = FALSE,
  exclude_soundtracks = FALSE,
  featured_tagids = NULL,
  rebuild_queue_if_stale = FALSE,
  ignore_user_preferences = FALSE,
  no_experimental_results = FALSE
)

get_discovery_settings(queue_type = NULL)

get_discovery_skipped(queue_type = NULL)
}
\arguments{
\item{queue_type}{Integer indicating the queue type. Possible values can
be retrieved using \code{node_enum("StoreDiscoveryQueueType")}.}

\item{country_code}{ISO 3166 country code representing the country from
which to view the Steam store.}

\item{rebuild_queue}{If \code{TRUE}, rebuilds the queue. Otherwise,
takes the previously generated one. Rebuilding the queue can lead to
error codes 2. In this case, retrying helps.}

\item{settings_changed}{Whether to overwrite the settings in
\code{get_discovery_settings} with the settings passed to
\code{get_discovery_queue}.}

\item{os_win}{If \code{TRUE}, only includes apps for Windows.}

\item{os_mac}{If \code{TRUE}, only includes apps for Mac OS.}

\item{os_linux}{If \code{TRUE}, only includes apps for Linux.}

\item{full_controller_support}{If \code{TRUE}, only includes apps with
full controller support.}

\item{native_steam_controller}{If \code{TRUE}, only includes apps with
native Steam controller support.}

\item{include_coming_soon}{If \code{TRUE}, includes apps that are coming
soon.}

\item{excluded_tagids}{A vector of tagIDs to exclude.}

\item{exclude_early_access, exclude_videos, exclude_software, exclude_dlc, exclude_soundtracks}{If \code{TRUE}, excludes early access apps, videos, software, DLCs, or
soundtracks.}

\item{featured_tagids}{A vector of tagIDs that must be present in the
output appIDs.}

\item{rebuild_queue_if_stale}{If \code{TRUE}, rebuilds the discovery queue
if it is stale. It is a mystery what stale means.}

\item{ignore_user_preferences}{If \code{TRUE}, ignores user preferences.}

\item{no_experimental_results}{Unknown.}
}
\value{
\describe{
 \item{\code{get_discovery_queue}}{A vector of appIDs in the requested
 discovery queue.}
 \item{\code{get_discovery_settings}}{A list of current discovery queue
 settings.}
 \item{\code{get_discovery_skipped}}{A vector of appIDs skipped in the
 discovery queue.}
}
}
\description{
Retrieve suggested and skipped apps on the discovery queue and its
settings.
}
\examples{
# return discovery queue only containing DLCs on sale
get_discovery_queue(queue_type = 8) # 8 - DLCOnSale

# return the best DLCs on sale in Germany
get_discovery_queue(queue_type = 8, country_code = "DE")

\dontrun{
# overwrite discovery queue settings
# NOTE: this modifies account settings and should be used with care
get_discovery_queue(settings_changed = TRUE, exclude_dlc = TRUE)
}
}
