% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievements.R
\name{get_game_achievements}
\alias{get_game_achievements}
\title{Achievements}
\usage{
get_game_achievements(appid, language = "english")
}
\arguments{
\item{appid}{appID of an application in the Steam store.}

\item{language}{Language string of the returned information.}
}
\value{
A dataframe containing information about a game's achievements.
}
\description{
Retrieve achievements for a game.
}
