% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_details.R
\name{get_game_details}
\alias{get_game_details}
\title{Create a Game Details S3 Object}
\usage{
get_game_details(app_id)
}
\arguments{
\item{app_id}{The Steam App ID of the game.}
}
\value{
An S3 object of class 'game_details'.
}
\description{
This function generates an S3 object summarizing a game's details.
}
\examples{
game <- get_game_details(72850)
print(game)
summary(game)
}
