% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_groups.R
\name{get_game_group}
\alias{get_game_group}
\title{Game groups}
\usage{
get_game_group(appids)
}
\arguments{
\item{appids}{A vector of multiple appIDs of applications in the Steam store.
There are generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query applications programmatically, e.g. using \code{\link{search_apps}}}
}}
}
\value{
A dataframe containing information about the game group
associated with an application. Notably, the columns \code{clanAccountID}
and \code{clanSteamIDString} contain the clanID and gid, respectively.
Each row represents one appID.
}
\description{
Retrieve details about the game groups (or community hubs) associated with an
appID.
}
\examples{
\dontrun{
# get the community hubs of Counter Strike and Team Fortress
get_game_group(c(10, 440))
}
}
\seealso{
\code{\link{resolve_vanity}}

\code{\link{parse_steam64}}
}
