% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-global-achievement-percentages-for-app.R
\name{get_global_achievement_percentages_for_app}
\alias{get_global_achievement_percentages_for_app}
\title{Get Global Achievement Percentages for App}
\usage{
get_global_achievement_percentages_for_app(app_id)
}
\arguments{
\item{app_id}{The app's AppID.}
}
\value{
The latest global achievement percentages for an app.
}
\description{
Get the global achievement percentages for an app.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{name}{The achievement's internal name}
  \item{percent}{The percentage of user's who have completed the achievement}
}
}
\examples{
# get global achievement percentages for Super Meat Boy
get_global_achievement_percentages_for_app(40800)
}
\seealso{
\code{\link{get_app_list}} to get a list of app names & id's.
}

