% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_browse.R
\name{get_hardware_items}
\alias{get_hardware_items}
\title{Get hardware items}
\usage{
get_hardware_items(
  items,
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L
)
}
\arguments{
\item{items}{Numeric vector of package IDs or a list of store item objects
as returned by \code{\link{store_items}(..., type = "packageid")}.}

\item{language}{Language string of the returned information.}

\item{elanguage}{Language identifier of the returned information.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{steam_realm}{Number describing the Steam realm.}
}
\value{
A dataframe containing information on hardware items.
}
\description{
Get information about items that require physical shipping.
}
\examples{
\dontrun{
  get_hardware_items(c(354231, 1628580))
  get_hardware_items(store_items(c(354231, 1628580), type = "packageid"))
}

}
\seealso{
Other store: 
\code{\link{get_items}()},
\code{\link{get_most_popular_tags}()},
\code{\link{get_store_categories}()},
\code{\link{store_item}()}
}
\concept{store}
