% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_json}
\alias{get_json}
\title{Get JSON from URL}
\usage{
get_json(url)
}
\arguments{
\item{url}{a string giving the URL}
}
\value{
An R object created by mapping the JSON content
  from the given URL to its R equivalent.
}
\description{
Get JSON-encoded data from a URL
}

