% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_info.R
\name{get_logon_info}
\alias{get_logon_info}
\alias{get_client_info}
\title{Session info}
\usage{
get_logon_info()

get_client_info(client_instanceid)
}
\arguments{
\item{client_instanceid}{InstanceID of an authenticated session. Can be
retrieved with \code{get_login_info}.}
}
\value{
\describe{
 \item{\code{get_logon_info}}{A dataframe where each row is an authenticated
 session. Contains columns on the client instanceID, protocol version,
 operating system, machine name, device type, and Steam realm.}
 \item{\code{get_client_info}}{A dataframe with a single row containing
 information on package version, operating system and client details.}
}
}
\description{
Retrieve login and client info of the authenticated session.
}
\examples{
\dontrun{
# authentication needed
auth_qr()

# get info about the session
info <- get_logon_info()
client <- info$client_instanceid

# get details on an instanceid
get_client_info(client)
}
}
