% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_service.R
\name{get_most_popular_tags}
\alias{get_most_popular_tags}
\title{Get tags}
\usage{
get_most_popular_tags(language = "english")
}
\arguments{
\item{language}{Language string of the returned information.}
}
\value{
A dataframe containing all tags sorted by their popularity
}
\description{
Get a list of whitelisted and localized tags sorted by their popularity.
}
\examples{
\dontrun{
get_most_popular_tags(language = "DE")
}

}
\seealso{
Other store: 
\code{\link{get_hardware_items}()},
\code{\link{get_items}()},
\code{\link{get_store_categories}()},
\code{\link{store_item}()}
}
\concept{store}
