% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{get_news}
\alias{get_news}
\title{News}
\usage{
get_news(
  appid,
  maxlength = NULL,
  enddate = NULL,
  count = 20,
  feeds = NULL,
  tags = NULL
)
}
\arguments{
\item{appid}{A single appID of an application in the Steam store.}
}
\description{
Get news data for a given application.
}
