% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-news-for-app.R
\name{get_news_for_app}
\alias{get_news_for_app}
\title{Get news for an app}
\usage{
get_news_for_app(app_id, count, max_length)
}
\arguments{
\item{app_id}{The app's AppID.}

\item{count}{The number of news entries to return.
Returns all news entries if unspecified.}

\item{max_length}{The maximum length of each news entry.
  Returns all characters if unspecified.}
}
\value{
The latest news of a game specified by its AppID.
}
\description{
Get the latest news for an app.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{gid}{The news entry's id}
  \item{title}{The news entry's title}
  \item{url}{The news entry's url}
  \item{author}{The news entry's author}
  \item{contents}{A shortened excerpt of the contents (to \code{max_length} characters),
     terminated by "..." if longer than \code{max_length}.}
  \item{date}{A UNIX timestamp}
}
}
\examples{
# get the most recent The Binding of Isaac: Rebirth news post
get_news_for_app(250900, count = 1)

# get all SteamWorld Dig news posts
get_news_for_app(252410)
}
\seealso{
\code{\link{get_app_list}} to get a list of app names & id's.
}

