% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opentarget.R
\name{get_opentarget}
\alias{get_opentarget}
\title{Process OpenTargets Data}
\usage{
get_opentarget(opentarget_dir, mouse_genes = NULL)
}
\arguments{
\item{opentarget_dir}{A string specifying the directory containing OpenTargets files.}

\item{mouse_genes}{Optional: Path to a file containing human-to-mouse gene conversion table.
The file should have columns `hgnc_symbol` and `mgi_symbol`.}
}
\value{
A list of data frames, where each data frame contains OpenTargets data for a specific disease.
The names of the list correspond to the disease names extracted from the file names.
}
\description{
This function processes OpenTargets data by reading disease-related files and optionally converting
human gene symbols to mouse gene symbols using a provided conversion table.
}
\examples{
\dontrun{
opentarget_dir <- "path/to/opentarget/files"
mouse_genes <- "path/to/mouse_genes.tsv"
OT.files <- get_opentarget(opentarget_dir, mouse_genes)
}
}
