% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-owned-games.R
\name{get_owned_games}
\alias{get_owned_games}
\title{Get a User's Owned Games}
\usage{
get_owned_games(steam_key, steam64_id, include_appinfo = TRUE,
  include_played_free_games = FALSE)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam64_id}{A 64-bit SteamID}

\item{include_appinfo}{Logical. Include game name and logo information in the output.}

\item{include_played_free_games}{Logical. By default, free games like Team Fortress 2 are excluded
If \code{include_played_free_games} is set to \code{TRUE}, they will be returned
if the player has played them at some point.}
}
\value{
A list of owned games
}
\description{
Returns a list of games a player owns along with some playtime information,
  if the profile is publicly visible. Private, friends-only,
  and other privacy settings are not supported.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{appid}{The unique identifier for the game}
  \item{name}{The name of the game}
  \item{playtime_2weeks}{The total number of minutes played in the last 2 weeks.}
  \item{playtime_forever}{The total number of minutes played on record,
     since Steam began tracking total playtime in early 2009.}
  \item{img_icon_url}{The filename for the game's icon.
     Accessible at \url{http://media.steampowered.com/steamcommunity/public/images/apps/APP_ID/IMG_ICON_URL.jpg}
     replacing \emph{APP_ID} and \emph{IMG_ICON_URL} as necessary.}
  \item{img_logo_url}{The filename for the game's logo.
     Accessible at \url{http://media.steampowered.com/steamcommunity/public/images/apps/APP_ID/IMG_LOGO_URL.jpg}
     replacing \emph{APP_ID} and \emph{IMG_LOGO_URL} as necessary.}
  \item{has_community_visible_stats}{Whether the program has stats accessible
     via \code{\link{get_user_stats_for_game}}}
}
}

