% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-player-bans.R
\name{get_player_bans}
\alias{get_player_bans}
\title{Get Player Bans}
\usage{
get_player_bans(steam_key, steam64_ids = c())
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam64_ids}{A vector of 64-bit Steam ID's.}
}
\value{
Steam Community, VAC, and Economy ban statuses for
  a list of Steam ID's.
}
\description{
Returns Steam Community, VAC, and Economy ban statuses for a list of Steam ID's.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{SteamId}{The player's 64 bit ID.}
  \item{CommunityBanned}{Indicates whether or not the player is banned
     from \href{http://steamcommunity.com/}{Steam Community}}
  \item{VACBanned}{Indicates whether or not the player has VAC bans on record.}
  \item{NumberOfVACBans}{Number of VAC bans}
  \item{DaysSinceLastBan}{Number of days since last ban}
  \item{EconomyBan}{The player's ban status in the economy.
     If the player has no bans on record the string will be "none",
     if the player is on probation it will say "probation", etc.}
}
}

