% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_player_summaries}
\alias{get_player_summaries}
\title{Get Player Summaries}
\usage{
get_player_summaries(steam_key, steam64_ids = c())
}
\arguments{
\item{steam_key}{a Steam Web API Key.
Get a key at \url{http://steamcommunity.com/dev/apikey}}

\item{steam64_ids}{a vector of 64-bit Steam ID's.
  Up to 100 Steam IDs can be requested.}
}
\value{
Basic profile information for a list of Steam ID's.

Some data associated with a Steam account may be hidden
if the user has their profile visibility set to "Friends Only"
or "Private". In that case, only public data will be returned.
}
\description{
Get basic profile information for a list of 64-bit Steam ID's.
}

