% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_player_summary}
\alias{get_player_summary}
\title{Get user info}
\usage{
get_player_summary(steamids)
}
\arguments{
\item{steamid}{Steam user ID as returned by \code{\link{get_steam_id}}.}

\item{relationship}{Type of relationship to the user. One of \code{all}
and \code{friend}.}
}
\value{
\itemize{
\item{\code{get_friend_list} returns a dataframe on relationship status
and time.}
\item{\code{\code{get_user_stats_for_game} returns a list with names of the
type of statistic, its value, and the type of statistic (achievement / stat).}}
\item{\code{get_user_group} returns a character vector of group IDs.}
\item{\code{get_player_summary returns a dataframe with one row for each
user containing basic information on the user account.}}
\item{\code{get_player_bans} returns a dataframe containing information
on bans (e.g. VAC, community bans)}
}
}
\description{
Functions to retrieve information on a given user including friends,
stats, achievements, groups.
}
\examples{
\dontrun{
# get steam ID
steamid <- get_steam_id("gabelogannewell")

# get all friends
get_friend_list(steamid)

# user stats for team fortress
get_user_stats_for_game
}

}
