% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_profile_items}
\alias{get_profile_items}
\title{Profiles}
\usage{
get_profile_items(
  steamid,
  language = "english",
  status = c("owned", "equipped")
)
}
\arguments{
\item{steamid}{SteamID of a user. The SteamID must be in a format that can
be converted by \code{\link{convert_steamid}}. This includes vanity,
Steam64, Steam2, and Steam3 IDs.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}
}
\description{
Fetch cosmetic information from a user profile.
}
