% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recently-played-games.R
\name{get_recently_played_games}
\alias{get_recently_played_games}
\title{Get Recently Played Games}
\usage{
get_recently_played_games(steam_key, steam64_id, count)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam64_id}{A 64-bit SteamID.}

\item{count}{The number of games to return. 
Returns all games if unspecified.}
}
\value{
A list of games a player has played in the last two weeks.
}
\description{
Returns returns a list of games a player has played 
in the last two weeks if the profile is publicly visible.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{appid}{The unique identifier for the game}
  \item{name}{The name of the game}
  \item{playtime_2weeks}{The total number of minutes played in the last 2 weeks.}
  \item{playtime_forever}{The total number of minutes played on record, 
     since Steam began tracking total playtime in early 2009.}
  \item{img_icon_url}{The filename for the game's icon. 
     Accessible at \url{http://media.steampowered.com/steamcommunity/public/images/apps/APP_ID/IMG_ICON_URL.jpg}
     replacing \emph{APP_ID} and \emph{IMG_ICON_URL} as necessary.}
  \item{img_logo_url}{The filename for the game's logo.
     Accessible at \url{http://media.steampowered.com/steamcommunity/public/images/apps/APP_ID/IMG_LOGO_URL.jpg}
     replacing \emph{APP_ID} and \emph{IMG_LOGO_URL} as necessary.}
}
}
\seealso{
\code{\link{get_user_stats_for_game}}
}

