% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-schema-for-game.R
\name{get_schema_achievements}
\alias{get_schema_achievements}
\title{Get Game's Achievements}
\usage{
get_schema_achievements(steam_key, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{app_id}{The app's AppID.}
}
\value{
A list of achievements available in a game.
}
\description{
Returns a list of achivements available in a game.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{name}{The name}
  \item{defaultvalue}{The default value}
  \item{displayName}{The display name}
  \item{hidden}{"0" if the achievement is hidden}
  \item{description}{A description}
  \item{icon}{The url for the achievement's "unlocked" image}
  \item{icongray}{The url for the achievement's "locked" image}
}
}
\seealso{
\code{\link{get_schema_for_game}}, \code{\link{get_schema_name}},
  \code{\link{get_schema_version}}, \code{\link{get_schema_stats}}
}

