% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-schema-for-game.R
\name{get_schema_for_game}
\alias{get_schema_for_game}
\title{Get Schema for Game}
\usage{
get_schema_for_game(steam_key, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{app_id}{The app's AppID.}
}
\value{
The game's schema (name, version, available achivements & stats)
}
\description{
Returns the Game's schema (name, version, available achivements & stats).
}
\details{
This is a wrapper for:

\code{\link{get_schema_name}}

\code{\link{get_schema_version}}

\code{\link{get_schema_achievements}}

\code{\link{get_schema_stats}}
}
\seealso{
\code{\link{get_schema_name}}, \code{\link{get_schema_version}}, 
  \code{\link{get_schema_achievements}}, \code{\link{get_schema_stats}}
}

