% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-schema-for-game.R
\name{get_schema_name}
\alias{get_schema_name}
\title{Get Game's Name}
\usage{
get_schema_name(steam_key, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{app_id}{The app's AppID.}
}
\value{
The game's name
}
\description{
Returns the game's name.
}
\section{Steam Web API Key}{

Get a Steam Web API Key at \url{http://steamcommunity.com/dev/apikey}
}
\seealso{
\code{\link{get_schema_for_game}}, \code{\link{get_schema_version}},
  \code{\link{get_schema_achievements}}, \code{\link{get_schema_stats}}
}

