% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-schema-for-game.R
\name{get_schema_stats}
\alias{get_schema_stats}
\title{Get Game Achivement Stats}
\usage{
get_schema_stats(steam_key, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{app_id}{The app's AppID.}
}
\value{
A list of achivement stats in a game.
}
\description{
Returns a list of achivement stats in a game.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{name}{The name}
  \item{defaultvalue}{The default value}
  \item{displayName}{The display name}
}
}
\seealso{
\code{\link{get_schema_for_game}}, \code{\link{get_schema_name}},
  \code{\link{get_schema_version}}, \code{\link{get_schema_achievements}}
}

