% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_steam_id}
\alias{get_steam_id}
\title{Get User ID}
\usage{
get_steam_id(name, type = "profile")
}
\arguments{
\item{name}{Name of the user account. This is not the display name,
but the account name. The name can be retrieved from the profile URL, i.e.
\code{https://steamcommunity.com/id/{name}/}.}

\item{type}{Type of Steam name. \code{profile} returns the Steam ID of
a user profile, \code{group} returns the ID of a public group and
\code{game_group} returns the ID of a game's official game hub.}
}
\value{
A length-1 character vector containing the Steam ID corresponding
to the input name.
}
\description{
Retrieves the Steam ID of a user profile, group or game hub.
}
\examples{
# get user ID
get_steam_id("gabelogannewell")

# get group ID
get_steam_id("SteamDB", type = "group")

# get game hub ID
get_steam_id("Break_the_Game", type = "game_group")


}
