% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-steam-id-64.R
\name{get_steam_id_64}
\alias{get_steam_id_64}
\title{Get a 64-bit Steam ID}
\usage{
get_steam_id_64(steam_key, steam_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam_id}{A Steam ID}
}
\value{
a 64-bit Steam ID
}
\description{
Get a 64-bit Steam ID
}

