% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_stats.R
\name{get_userdata}
\alias{get_userdata}
\alias{get_app_user_details}
\alias{get_library_stats}
\alias{get_owned_apps}
\alias{get_badge}
\alias{get_badge_info}
\alias{recommend_apps}
\title{User data}
\usage{
get_userdata()

get_app_user_details(appids)

get_library_stats()

get_owned_apps()

get_badge(appid, badgeid)

get_badge_info(appid)

recommend_apps(appid)
}
\description{
Authenticated functions to fetch information about the authenticated user.
}
