% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishlist.R
\name{get_wishlist}
\alias{get_wishlist}
\title{Wishlist}
\usage{
get_wishlist(steamid)
}
\arguments{
\item{steamid}{SteamID of a user. The SteamID must be in a format that can
be converted by \code{\link{convert_steamid}}. This includes vanity,
Steam64, Steam2, and Steam3 IDs.}
}
\value{
A list of apps where each field contains information about
name, reviews, release dates, platforms, screenshots, discounts,
tags, and compatibility.
}
\description{
Get wishlist details of a user.
}
\examples{
get_wishlist("76561198092541763")

}
