% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glblachievementpct.R
\name{glbachieve}
\alias{glbachieve}
\title{Global achievement percentage for a game}
\usage{
glbachieve(gameid = "440")
}
\arguments{
\item{gameid}{Steam game ID}
}
\description{
Global achievement percentage for a game
}

