% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstatsforgame.R
\name{glbstats}
\alias{glbstats}
\title{Global stats for a game}
\usage{
glbstats(gameid = "17740", stat = "global.map.emp_isle")
}
\arguments{
\item{gameid}{Steam game ID}

\item{stat}{name of achievent}
}
\description{
Returns list for gloabl stats for a game
}

