% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnews.R
\name{gnews}
\alias{gnews}
\title{Latest news for a game}
\usage{
gnews(gameid = "440", count = "3")
}
\arguments{
\item{gameid}{Steam game ID}

\item{count}{number of news entries to return}
}
\description{
Data frame for steam game news
}

