% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcode.R
\name{html_to_bbcode}
\alias{html_to_bbcode}
\title{HTML to BBCode}
\usage{
html_to_bbcode(x, preserve_newlines = FALSE)
}
\arguments{
\item{x}{A string containing HTML formatting.}

\item{preserve_newlines}{If \code{TRUE}, preserves linebreaks. If
\code{FALSE}, also preserves linebreaks -- or so it seems.}
}
\value{
A character string containing BBCode formatted text.
}
\description{
Converts an HTML string to BBCode. BBCode is commonly used in the Steam
forum and other bulletin boards.
}
\examples{
string1 <- "<b>test</b>"
string2 <- "<br>\n"
string3 <- "<a href='google.com'>link</a>"

html_to_bbcode(string1)
html_to_bbcode(string2)
html_to_bbcode(string3)
}
