% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is-playing-shared-game.R
\name{is_playing_shared_game}
\alias{is_playing_shared_game}
\title{Is Playing Shared Game}
\usage{
is_playing_shared_game(steam_key, steam64_id, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam64_id}{A 64-bit SteamID.}

\item{app_id}{The app's AppID.}
}
\value{
\code{TRUE} if a borrowing account is currently playing this game.
  Returns \code{FALSE} if the game is not borrowed or the borrower currently doesn't play this game.
}
\description{
Returns \code{TRUE} if a borrowing account is currently playing this game.
Returns \code{FALSE} if the game is not borrowed or the borrower currently doesn't play this game.
}

