% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamid.R
\name{is_steam64}
\alias{is_steam64}
\alias{is_steam2}
\alias{is_steam3}
\alias{is_vanity}
\title{Steam ID checks}
\usage{
is_steam64(x)

is_steam2(x)

is_steam3(x)

is_vanity(x)
}
\arguments{
\item{x}{A character string that is to be checked.}
}
\description{
Checks if a given string can be parsed as a Steam ID. Functions are defined
for ID64, Steam2, Steam3, and vanity IDs.
}
\details{
\code{is_steam64} tests if a character string is a 64-bit representation
of a decimal number.
\code{is_steam2} and \code{is_steam3} test if \code{x} follows the
string format of textual Steam ID representations as described
\href{https://developer.valvesoftware.com/wiki/SteamID#As_Represented_Textually}{here}.
\code{is_vanity} returns \code{TRUE}, if all other functions return
\code{FALSE}.
}
\examples{
\dontrun{
is_steam64("76561197984981409", "12345") # TRUE, FALSE
is_steam2("STEAM_1:1:12357840", "STEAM_5:6:123") # TRUE, FALSE
is_steam3("[U:1:24715681]", "U:4:1234") # TRUE, FALSE
is_vanity("12345", "76561197984981409") # TRUE, FALSE
}
}
