% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spatial_data.R
\name{load_spatial_data}
\alias{load_spatial_data}
\title{Load Spatial Data}
\usage{
load_spatial_data(
  infoTable,
  minUMICountsPerGene = 100,
  minSpotsPerGene = 5,
  minUMICountsPerSpot = 500,
  platform = "Visium"
)
}
\arguments{
\item{infoTable}{Path to the CSV file containing information about spatial data files.}

\item{minUMICountsPerGene}{Minimum UMI counts per gene for filtering.}

\item{minSpotsPerGene}{Minimum spots per gene for filtering.}

\item{minUMICountsPerSpot}{Minimum UMI counts per spot for filtering.}

\item{platform}{Platform used for spatial transcriptomics (default is "Visium").}
}
\value{
A Seurat object containing the spatial transcriptomics data.
}
\description{
Load spatial transcriptomics data from an infoTable file.
}
\examples{
\dontrun{
  se <- load_spatial_data(
    infoTable = "path/to/infoTable.csv",
    minUMICountsPerGene = 100, 
    minSpotsPerGene = 5,
    minUMICountsPerSpot = 500,
    platform = "Visium"
  )
}
}
