% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{login}
\alias{login}
\title{Login}
\usage{
login()
}
\value{
Error code returned by \code{\link{system}}.
}
\description{
Opens the Steam login page in the default browser. A Steam login is required
for many of the unofficial API methods.
}
\examples{
\dontrun{
login()
}
}
