% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{content_descriptors}
\alias{content_descriptors}
\alias{universes}
\alias{account_types}
\alias{friend_relationships}
\alias{steam_currencies}
\alias{published_file_query_types}
\alias{user_badges}
\alias{parse_steamkit_enum}
\title{Enums}
\usage{
content_descriptors()

universes()

account_types()

friend_relationships()

steam_currencies()

published_file_query_types()

user_badges()

parse_steamkit_enum(
  enum = NULL,
  type = c("Enums", "EnumsProductInfo", "SteamLanguage")
)
}
\arguments{
\item{enum}{Name of the enumeration to retrieve. If \code{NULL}, returns
the names of available enums.}

\item{type}{Type of enums. Corresponds to a generated file from
SteamKit.}
}
\value{
If \code{enum} is not \code{NULL}, returns a dataframe
containing code and descriptions. Otherwise, returns the names of
available enums.
}
\description{
The Steam API defines a number of enums for use with the API. These enums
can help resolve some common identifiers such as types and categories.
The enums are scraped from the C# library
\href{https://github.com/SteamRE/SteamKit}{SteamKit}, which
auto-generates enums from protobufs.

Some enums are pre-defined, either internally or as exported functions.
Non-defined enums can be retrieved using \code{parse_steamkit_enum}.
}
\examples{
# investigate all enums of SteamLanguage.cs
parse_steamkit_enum(type = "SteamLanguage")

# retrieve all code descriptions for SteamRealm
# Steam realms can be passed as arguments in various store endpoints
parse_steamkit_enum("SteamRealm", type = "SteamLanguage")

# retrieve universe descriptions
# can be useful to interpret output of steamid parsing
parse_steamkit_enum("Universe", type = "SteamLanguage")

# retrieve communty item classes
# useful to decipher the output of functions like query_loyalty_rewards
parse_steamkit_enum("CommunityItemClass", type = "Enums")
}
