% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_null_matrix.R
\name{permutation_null_matrix}
\alias{permutation_null_matrix}
\title{Generate a Permutation Null Matrix for Spatial Data}
\usage{
permutation_null_matrix(
  info_table_path,
  seurat_clusters_path,
  gnum_path,
  output_dir,
  idx = 1,
  min_umi_gene = 100,
  min_spots_gene = 5,
  min_umi_spot = 500,
  platform = "Visium",
  num_permutations = 10000,
  workers = 10,
  max_globals_size = 2 * 1024^3
)
}
\arguments{
\item{info_table_path}{Path to the infoTable CSV file containing spatial data information.}

\item{seurat_clusters_path}{Path to the CSV file containing cluster annotations.}

\item{gnum_path}{Path to the file containing the number of genes in the NULL set.}

\item{output_dir}{Directory to save the resulting permutation matrix.}

\item{idx}{Index to select the specific gene set size from gnum.}

\item{min_umi_gene}{Minimum UMI counts per gene (default: 100).}

\item{min_spots_gene}{Minimum number of spots per gene (default: 5).}

\item{min_umi_spot}{Minimum UMI counts per spot (default: 500).}

\item{platform}{Sequencing platform (default: "Visium").}

\item{num_permutations}{Number of permutations (default: 10000).}

\item{workers}{Number of parallel workers (default: 10).}

\item{max_globals_size}{Maximum global size for parallelization (default: 2 GB).}
}
\value{
A list containing the permutation matrix and execution time.
}
\description{
This function calculates a permutation null matrix by randomizing genes across clusters.
}
\examples{
\dontrun{
permutation_null_matrix(
  info_table_path = "path/to/infoTable.csv",
  seurat_clusters_path = "path/to/seurat_clusters.csv",
  gnum_path = "path/to/gnum.txt",
  output_dir = "output/directory",
  idx = 3
)
}
}
