% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_dist.R
\name{plot_score_dist}
\alias{plot_score_dist}
\title{Plot Enrichment Score Distributions for Clusters}
\usage{
plot_score_dist(
  se,
  perm.mat,
  cluster_anno = "seurat_clusters",
  clusters_to_plot = NULL,
  enrichment_score_col
)
}
\arguments{
\item{se}{A Seurat object containing enrichment scores in the metadata.}

\item{perm.mat}{A null distribution permutation matrix of enrichment scores per cluster (global medians).
The permutation matrix is generated using the generatePermutationMatrix function and must be generated
from a gene list of equal size as the gene list score given in 'enrichment_score_col'.}

\item{cluster_anno}{The column in the Seurat metadata that contains cluster labels (default is "seurat_clusters").}

\item{clusters_to_plot}{Optional vector of indices to subset and plot specific clusters only.}

\item{enrichment_score_col}{A string specifying the enrichment score column to use.}
}
\value{
No return value. Produces a series of ggplot2 density plots, one per cluster.
}
\description{
Generates density plots of null distribution scores for each cluster
and overlays the actual enrichment score to visually compare significance.
This function is useful for assessing whether observed enrichment scores
deviate from a random background.
}
\examples{
plotScoreDist(
  se = se,
  perm.mat = perm.mat.genetic.data,
  cluster_anno = "supercluster_term",
  enrichment_score_col = "OpenTargets_SCZ_Genetic_1"
)
}
