% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_details.R
\name{print.game_details}
\alias{print.game_details}
\title{Prints Game Details}
\usage{
\method{print}{game_details}(x, ...)
}
\arguments{
\item{x}{game_details Object}

\item{...}{ignored}
}
\description{
Prints Game Details
}
