% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_save_opentarget.R
\name{process_and_save_opentarget}
\alias{process_and_save_opentarget}
\title{Process and Save OpenTarget Gene Sets and Windows}
\usage{
process_and_save_opentarget(
  all_gene_sets,
  windows_list,
  conditions,
  base_gene_set_dir,
  base_window_list_dir,
  species = "mouse"
)
}
\arguments{
\item{all_gene_sets}{A list of all gene sets.}

\item{windows_list}{A list of overlapping windows for each gene set.}

\item{conditions}{A list of conditions with their respective gene set indices (e.g., `list(list("EPI", 2), ...)`).}

\item{base_gene_set_dir}{Directory to save processed gene sets.}

\item{base_window_list_dir}{Directory to save overlapping windows.}

\item{species}{Species name for file naming (e.g., "mouse").}
}
\description{
This function processes OpenTarget gene sets for multiple conditions and saves the results.
}
\examples{
\dontrun{
process_and_save_opentarget(
  all_gene_sets = all_gene_sets,
  windows_list = windows_list,
  conditions = list(list("EPI", 2), list("ASD", 5)),
  base_gene_set_dir = "path/to/gene_sets",
  base_window_list_dir = "path/to/windows",
  species = "mouse"
)
}
}
