% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{query_market_assets}
\alias{query_market_assets}
\alias{get_asset_prices}
\alias{get_asset_info}
\alias{get_price_history}
\title{Assets}
\usage{
query_market_assets(
  appid,
  search_term = NULL,
  sort_column = NULL,
  sort_dir = NULL,
  search_description = NULL,
  paginate = FALSE,
  max_pages = Inf
)

get_asset_prices(appid, currency = NULL, language = "english")

get_asset_info(
  appid,
  classids,
  class_count = 1L,
  language = "english",
  instanceids = NULL
)

get_price_history(appid, market_hash_name)
}
\arguments{
\item{appid}{A single appID of an application in the Steam store.}

\item{search_term}{Term to search for in the market item name.}

\item{sort_column}{Column to sort the response by.}

\item{sort_dir}{Direction of sorting if \code{sort_column} is not
\code{NULL}. Can be \code{asc} or \code{desc}.}

\item{paginate}{If \code{TRUE}, paginates through the results. Some methods
provide data access in digestible chunks and do not return all available
information at once. The \code{paginate} automatically retrieves all pages
until \code{max_pages} is reached.}

\item{max_pages}{Maximum number of pages to paginate. Ignored if
\code{paginate} is \code{FALSE}. Defaults to \code{Inf} such that all
available data are paginated.}

\item{currency}{ISO 4217 code to filter by currency. Defaults to all
available currencies.}

\item{classids}{A vector of classIDs of the assets to retrieve.}

\item{instanceids}{A vector of instanceIDs corresponding to the}

\item{search_descriptions}{Unknown.}
}
\value{
\describe{
 \item{\code{query_market_assets}}{A dataframe with each row representing
 one asset. The dataframe provides data on name, hash names, prices,
 the corresponding app, asset classIDs, and the output of
 \code{get_asset_info}.}

 \item{\code{get_asset_prices}}{A dataframe containg information on assert
 prices. \code{name}, \code{date}, \code{class}, and \code{classid} provide
 metadata about assets. Optionally, tags are provided in \code{tags} and
 \code{tag_ids}. \code{class}, \code{tags}, and \code{tag_ids} are nested
 as dataframes or lists as they can contain values of arbitrary length.
 \code{prices.\*} contain the current prices for
 the currencies specified in \code{currency}. If an asset is on sale,
 \code{original_prices.\*} show the original prices.}

 \item{\code{get_asset_info}}{A dataframe with each row representing
 asset info on each provided classID. \code{\*_url} columns show
 extensions to the URL \url{http://cdn.steamcommunity.com/economy/image/}.
 \code{name}, \code{market_hash_name} and \code{market_name} show various
 types of asset names. If an asset has received a fraud warning, the
 column \code{fraudwarnings} can contain their warning strings. The
 \code{descriptions} columns contain information about asset attributes,
 \code{actions} about action URLs, and \code{tags} about asset tags.}
}
}
\description{
Query market assets and their prices. An asset is an item
from the Steam economy that can be traded. For more information about
assets, see the \href{https://partner.steamgames.com/doc/store/assets}{Steamworks documentation}.
}
\examples{
# show asset prices for Team Fortress and all available currencies
get_asset_prices(440)

# show the same assets, but return only EUR
get_asset_prices(440, currency = "EUR")

# a list of valid currency codes can be retrieved using
# steamkit_enum
steamkit_enum("CurrencyCode", type = "SteamLanguage")

# get_asset_prices returns classIDs for each asset. These classIDs can
# be used to retrieve detailed information on an asset.
get_asset_class(730, c("5710093913", "5189384166"))
}
\references{
\url{https://wiki.teamfortress.com/wiki/WebAPI/GetAssetPrices}

\url{https://wiki.teamfortress.com/wiki/WebAPI/GetAssetClassInfo}
}
\seealso{
\code{\link{steamkit_enum}} to retrieve all currency codes

\code{\link{query_loyalty_rewards}} for loyalty assets
}
