% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{resolve_bundles}
\alias{resolve_bundles}
\alias{resolve_packages}
\title{Bundles and packages}
\usage{
resolve_bundles(bundleids, country_code = "US", language = "english")

resolve_packages(packageids, country_code = "US", language = "english")
}
\arguments{
\item{bundleids}{A vector of up to 200 bundleIDs.}

\item{packageids}{A vector of up to 200 packageIDs. For
\code{package_details}, this vector can be of any length.}
}
\value{
Both \code{resolve_bundles} and \code{resolve_packages} return
detailed information on a bundle or a package including prices,
release dates, compatibility information, discounts, and localization.
}
\description{
Resolve details of bundles and packages.
}
\examples{
# resolve Graveyard Keeper bundle
resolve_bundles(8077)

# resolve OST from the Graveyard Keeper bundle
resolve_packages(295588)

# get more details on OST
package_details(295588)
}
