% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_gene_sets.R
\name{score_gene_sets}
\alias{score_gene_sets}
\title{Score Gene Sets and Add Module Scores}
\usage{
score_gene_sets(se, gene_sets, ctrl = 100)
}
\arguments{
\item{se}{A Seurat object containing the spatial transcriptomics data.}

\item{gene_sets}{A list of gene sets, where each element is a character vector of gene names.}

\item{ctrl}{The number of control genes to use in the module score calculation (default: 100).}
}
\value{
A list containing:
  - `se`: The updated Seurat object with module scores.
  - `score_matrix`: A matrix with the percentage of genes present for each gene set.
}
\description{
This function calculates the percentage of genes from each gene set present in the Seurat object
and adds module scores for each gene set to the object.
}
