% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_apps}
\alias{search_apps}
\alias{suggest}
\alias{store_search}
\title{Search apps}
\usage{
search_apps(term, links = FALSE)

suggest(
  term,
  country_code = "US",
  language = "english",
  steam_realm = 1L,
  require_type = NULL,
  use_store_query = TRUE,
  use_search_spellcheck = TRUE,
  search_creators_and_tags = TRUE,
  links = FALSE
)

store_search(term, country_code = "US", language = "english")
}
\arguments{
\item{term}{Search term used to query the Steam store.}

\item{links}{Whether to include links to icon and logo in the output.
Defaults to \code{FALSE} to avoid littering the output.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{language}{Language string of the returned information.}

\item{steam_realm}{Number describing the Steam realm.}

\item{require_type}{Which application types should be returned? Can be
one of \code{bundle}, \code{game}, \code{demo}, \code{mod} \code{dlc},
\code{software}, \code{video}, \code{hardware}, \code{series},
and \code{music}.}

\item{use_store_query}{Unknown parameter.}

\item{use_search_spellcheck}{Whether search terms should be spellchecked.}

\item{search_creators_and_tags}{Whether to search for the search term in
creator names and tags.}
}
\value{
A dataframe containing the appid and application name and
(optionally) links to application icon and logo.
}
\description{
Simple Steam store queries that return the best matches to a given search
query. Uses Steam's unofficial API.

Steam provides different endpoints that essentially do the same, but
in a slightly different manner.
\itemize{
 \item{\code{search_apps} has less filtering capabilities, provides no
 context and returns a slim output.}
 \item{\code{suggest} provides many filtering capabilities, provides context
 and returns a slightly more detailed output.}
 \item{\code{store_search} has no filtering capabilities, provides context
 and returns a detailed output.}
}
}
\examples{
\dontrun{
# returns a single result
search_apps("team fortress")

# returns multiple results
search_apps("team")

# returns the first 10 results
suggest("team")

# returns the first 10 most fitting DLCs
suggest("team", require_type = "dlc")

# translates results to German and converts dollar to euro
suggest("team", language = "german", country_code = "DE")

# provides detailed data
store_search("team")
}
}
