% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{search_support}
\alias{search_support}
\title{Support FAQ}
\usage{
search_support(search_text, elanguage = NULL, paginate = TRUE, max_pages = Inf)
}
\arguments{
\item{search_text}{Search text to match FAQ articles by.}

\item{paginate}{If \code{TRUE}, paginates through the results. Some methods
provide data access in digestible chunks and do not return all available
information at once. The \code{paginate} automatically retrieves all pages
until \code{max_pages} is reached.}

\item{max_pages}{Maximum number of pages to paginate. Ignored if
\code{paginate} is \code{FALSE}. Defaults to \code{Inf} such that all
available data are paginated.}
}
\description{
Search support FAQ articles.
}
\examples{
# get help articles on portal 2
search_support("portal 2")

# get help articles on purchases
search_support("purchase")

# get localized articles in french
search_support("purchase", elanguage = 2)
}
