% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesfunctions.R
\name{separatebygenre}
\alias{separatebygenre}
\title{Function to separate rows with multiple genres into multiple rows}
\usage{
separatebygenre(games)
}
\arguments{
\item{games}{The "games" dataset}
}
\value{
games The "games" dataset with rows separated by genres
}
\description{
Function to separate rows with multiple genres into multiple rows
}
\keyword{internal}
