% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_permutation_matrix.R
\name{sort_cluster_names}
\alias{sort_cluster_names}
\title{Sort cluster names}
\usage{
sort_cluster_names(cluster_names)
}
\arguments{
\item{cluster_names}{A vector containing cluster identifiers.}
}
\value{
A vector of sorted cluster identifiers.
}
\description{
Sorts cluster identifiers by placing numeric cluster identifiers first
in numerical order, followed by non-numeric cluster identifiers in
alphabetical order.
}
