% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_with_overlap.R
\name{split_with_overlap}
\alias{split_with_overlap}
\title{Split a Vector into Overlapping Segments}
\usage{
split_with_overlap(vec, seg_length, step_size)
}
\arguments{
\item{vec}{A numeric or character vector to split.}

\item{seg_length}{The length of each segment.}

\item{step_size}{The step size for moving the sliding window.}
}
\value{
A list of overlapping segments from the input vector.
}
\description{
This function splits a vector into overlapping segments based on specified segment length and step size.
}
\examples{
\dontrun{
vec <- 1:10
result <- split_with_overlap(vec, seg_length = 4, step_size = 2)
print(result)
}
}
