% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{steam_api_args}
\alias{ARG_APP_ID}
\alias{ARG_COUNT}
\alias{ARG_FORMAT}
\alias{ARG_GAME_ID}
\alias{ARG_MAX_LENGTH}
\alias{steam_api_args}
\title{Steam Web API service definition argument constants}
\format{\preformatted{ chr "?appid="
}}
\usage{
ARG_APP_ID

ARG_GAME_ID

ARG_MAX_LENGTH

ARG_COUNT

ARG_FORMAT
}
\description{
Service definition argument constants for Steam Web API
}
\keyword{datasets}

