% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{steam_api_service_definitions}
\alias{STEAM_API_APP_LIST}
\alias{STEAM_API_GLOBAL_ACHIEVEMENT}
\alias{STEAM_API_NEWS}
\alias{STEAM_API_PLAYER_SUMMARIES}
\alias{STEAM_FRIENDS_LIST}
\alias{STEAM_RESOLVE_VANITY_URL}
\alias{steam_api_service_definitions}
\title{Steam Web API service definition constants}
\format{a character string}
\source{
\url{https://developer.valvesoftware.com/wiki/Steam_Web_API}
}
\usage{
STEAM_API_APP_LIST

STEAM_API_NEWS

STEAM_API_GLOBAL_ACHIEVEMENT

STEAM_API_PLAYER_SUMMARIES

STEAM_RESOLVE_VANITY_URL

STEAM_FRIENDS_LIST
}
\description{
Service definition constants for Steam Web API.
}
\seealso{
\code{\link{steam_api_args}} for service definition
  argument constants.
}
\keyword{datasets}

