% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam_game_data.r
\name{steam_game_data}
\alias{steam_game_data}
\title{Get Steam Game Data}
\usage{
steam_game_data(game_id, country = "us")
}
\arguments{
\item{game_id}{Steam game id}

\item{country}{Alpha-2 country code}
}
\description{
Steam data about a specific game
}
