% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{steam_stats}
\alias{steam_stats}
\alias{get_supported_api_list}
\alias{get_servertime}
\title{Meta info}
\usage{
steam_stats()

get_supported_api_list()

get_servertime()
}
\value{
\code{get_supported_api_list()} returns an unnamed list where each
index contains a named list with information on the respective API endpoint.
\code{get_servertime()} returns a \code{POSIXct} value. \code{steam_stats()}
contains a named list of length 2.
}
\description{
Fetch metadata about the Steam server.

\code{get_supported_api_list()} returns information about all supported
API endpoints of the web API. A complete list including undocumented
endpoints can be found on \url{https://steamapi.xpaw.me/}.

\code{get_servertime()} returns the current Steam server time.

\code{steam_stats()} returns the current users online and ingame.
}
\examples{
\dontrun{
get_supported_api_list()

get_servertime()

steam_stats()
}
}
