% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamr-package.R
\docType{package}
\name{steamr-package}
\alias{steamr}
\alias{steamr-package}
\title{steamr: R Interface to Steam's APIs}
\description{
R interface to the Steam Web API and internal API. Performs standardized API calls, session authentication, key management, Steam ID conversion and response formatting. Provides tools for data scraping of game details, user reviews, store economy data, news, and many more.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org}

}
\keyword{internal}
