% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{steamR}
\alias{steamR}
\alias{steamR-package}
\title{steamR}
\description{
An interface to Steam's Web API
}
\details{
This package provides a wrapper for Steam's Web API.
}
\author{
Jose Gallegos, \email{josegallegos07@gmail.com}
}
\seealso{
\url{http://steamcommunity.com/dev/apikey} to get a Steam Web API Key.
}

