% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR-package.R
\name{steamR_methods}
\alias{steamR_methods}
\title{Methods}
\usage{
steamR_methods()
}
\value{
A list of methods available in the steamR package.
}
\description{
Returns a list of available steamR methods.
}

