% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontpage.R
\name{stf_frontpage}
\alias{stf_frontpage}
\alias{stf_frontpage_category}
\alias{stf_frontpage_genre}
\title{Frontpage apps}
\usage{
stf_frontpage(
  language = "english",
  country_code = "US",
  trailer = NULL,
  extra = NULL
)

stf_frontpage_category(category, language = "english", country_code = "US")

stf_frontpage_genre(genre, language = "english", country_code = "US")
}
\arguments{
\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{country_code}{ISO 3166 country code representing the country from
which to view the Steam store. A list of Steam countries can be retrieved
with \code{\link{get_country_list}}.}

\item{category}{A category of store items as shown on the front page of the
store. Sensible values can be retrieved using \code{\link{get_categories}}.}

\item{genre}{A genre of store items as shown on the front page of the store.
Sensible values can be retrieved using \code{\link{get_genres}}.}
}
\value{
\code{stf_frontpage} returns a grouped dataframe where the
\code{category} column represents the category of apps on the frontpage,
e.g. \code{"cat_spotlight"} contains the current special offers.

\code{stf_frontpage_genre} and \code{stf_frontpage_category} return a
dataframe containing the tab category, an unknown type, and the ID of
the featured store item.
}
\description{
These functions return apps as they are currently shown on the (genre)
dashboard of the store front page (\url{https://store.steampowered.com/}).

\itemize{
 \item{\code{stf_frontpage} returns the featured apps on the main store
 front page.}
 \item{\code{stf_frontpage_category} returns the featured apps on a
 category-specific front page (e.g. new releases or top sellers).}
 \item{\code{stf_frontpage_genre} returns the featured apps on a
 genre-specific front page (e.g. action or racing).}
}
}
\examples{
\donttest{# get details about the daily deal in the swedish store
fp <- stf_frontpage(country_code = "SE")
fp[fp$category \%in\% "cat_dailydeal", ]

# get details about the top racing games
fp <- stf_frontpage_genre("racing")
fp <- fp[fp$tab \%in\% "topsellers", ]
wba_store_items(fp$id)}
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{stf_frontpage}          \tab no \tab no\cr
    \code{stf_frontpage_category} \tab no \tab no\cr
    \code{stf_frontpage_genre}    \tab no \tab no
}




}
