% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{stf_locations}
\alias{stf_locations}
\alias{wba_countries}
\title{Location}
\usage{
stf_locations(country = NULL, region = NULL)

wba_countries(language = "english")
}
\arguments{
\item{country}{ISO-2 country code. If specified, returns information
on regions, otherwise countries.}

\item{region}{Region code within a country specified in \code{country}.
Country must not be \code{NULL} if \code{region} is specified. If specified,
returns information on cities, otherwise on regions.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}
}
\value{
\describe{
 \item{\code{query_locations}}{A dataframe with country code, name,
 and whether the location can be further subdivided into states or cities.}

 \item{\code{get_country_list}}{A dataframe containing country code
 and country name.}
}
}
\description{
Retrieves information on Steam's country, region, and city codes.

\code{stf_locations} queries the storefront API and can drill down all
Steam location levels, while \code{wba_countries} can only provide
information on the country level.
}
\examples{
\donttest{# get country codes
stf_locations()

# get region codes of Germany
stf_locations(country = "DE")

# get city codes of Baden-Wurttemberg, Germany
stf__locations(country = "DE", region = "01")

# get localized country names
wba_countries(language = "german")}
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{stf_locations} \tab no \tab no\cr
    \code{wba_countries} \tab no \tab no
}




}
