% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievments.R
\name{stmachieve}
\alias{stmachieve}
\title{Steam achievments}
\usage{
stmachieve(key, id = "76561197973625640")
}
\arguments{
\item{key}{Steam API key}

\item{id}{Steam user ID}
}
\description{
get steam achievment list
}

