% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{store_data_request}
\alias{store_data_request}
\alias{store_context}
\title{Steam utility objects}
\usage{
store_data_request(include = NULL, apply_user_filters = FALSE)

store_context(
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L
)
}
\arguments{
\item{include}{List of extra information to include. Can be one or several
of the following: \code{release}, \code{platforms},
\code{all_purchase_options}, \code{screenshots}, \code{trailers},
\code{ratings}, \code{tag_count}, \code{reviews}, \code{basic_info},
\code{supported_languages}, \code{full_description}, \code{included_items}
and \code{assets_without_overrides}. If \code{all}, returns all information.}

\item{apply_user_filters}{Unknown.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{elanguage}{Numeric code representing the store language. A list
of language codes and their corresponding languages is defined in
\code{\link{ELanguage}}.}

\item{country_code}{ISO 3166 country code representing the country from
which to view the Steam store. A list of Steam countries can be retrieved
with \code{\link{get_country_list}}.}

\item{steam_realm}{Number describing the Steam realm. A value of 1
indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.}
}
\value{
\code{store_data_request()} returns a list of class
\code{StoreBrowseItemDataRequest}. \code{store_context()} returns a list of
class \code{StoreBrowseContext}.
}
\description{
Create utility objects commonly used in the Steam web API.

\itemize{
 \item{\code{store_data_request} specifies which extra information should
 be included in the API response. It is usually supplied over an argument
 \code{data_request}.}
 \item{\code{store_context} specifies the geographic and language context
 from which to access the Steam store. It is usually supplied over an
 argument \code{context}.}
}
}
\examples{
\donttest{# Specify the geographic context
wba_store_items(440, context = store_context(language = "german", country_code = "DE"))

# Specify the data to be returned
wba_store_items(440, data_request = store_data_request(c("screenshots", "trailers")))}
}
\seealso{
\code{\link{wba_store_items}}, \code{\link{wba_hardware_items}},
\code{\link{wba_charts_weekly}}, \code{\link{wba_query}},
\code{\link{wba_marketing_message}}
}
