% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_browse.R
\name{store_item}
\alias{store_item}
\alias{store_items}
\title{Define a store item}
\usage{
store_item(
  appid = NULL,
  packageid = NULL,
  bundleid = NULL,
  tagid = NULL,
  creatorid = NULL,
  hubcategoryid = NULL
)

store_items(items, type = NULL)
}
\arguments{
\item{appid}{Application ID}

\item{packageid}{Package ID}

\item{bundleid}{Bundle ID}

\item{tagid}{Tag ID}

\item{creatorid}{Creator ID}

\item{hubcategoryid}{Hub category ID}

\item{items}{Numeric vector of Steam IDs. Will be converted to
\code{store_item}s. If a list is passed, will return \code{items} as-is.}

\item{type}{ID type of all items to be returned. Must be one of the
argument names of \code{store_item}.}
}
\value{
An object of class \code{StoreItemID}.
}
\description{
Create an object describing a store item by ID. Can be an application,
package, bundle, tag, creator or hub category.

\code{store_item} returns a single ID object, but allows the passing
of varying ID types. \code{store_items} returns a list of ID objects, but
is restricted to a single ID type.
}
\examples{
store_item(10)

\dontrun{
ids <- list(
  store_item(268770),
  store_item(1182620),
  store_item(creatorid = 10)
)

get_items(ids)
}
}
\seealso{
Other store: 
\code{\link{get_hardware_items}()},
\code{\link{get_items}()},
\code{\link{get_store_categories}()}
}
\concept{store}
