% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesfunctions.R
\name{str_arr_to_vector}
\alias{str_arr_to_vector}
\title{Function to manually parse a serialized str array into a R vector.
If the array has no (valid) elements, NULL is returned.
If the string s is a non-array sequence, the sequence is returned as a single element}
\usage{
str_arr_to_vector(s)
}
\arguments{
\item{s}{a string containing an array like ['A', 'B']}
}
\value{
spl The vector of elements
}
\description{
Function to manually parse a serialized str array into a R vector.
If the array has no (valid) elements, NULL is returned.
If the string s is a non-array sequence, the sequence is returned as a single element
}
