% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_spatial_data.R
\name{subset_spatial_data}
\alias{subset_spatial_data}
\title{Subset Spatial Data by Cluster}
\usage{
subset_spatial_data(se, clusters)
}
\arguments{
\item{se}{The spatial data object (Seurat object).}

\item{clusters}{A vector of cluster IDs to subset.}
}
\value{
A named list of subsetted Seurat objects.
}
\description{
This function subsets the spatial data object (`se`) by cluster.
}
