% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_details.R
\name{summary.game_details}
\alias{summary.game_details}
\title{Print a Summary of Game Details}
\usage{
\method{summary}{game_details}(object, ...)
}
\arguments{
\item{object}{game_details Object}

\item{...}{ignored}
}
\description{
Print a Summary of Game Details
}
