% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_most_popular_tags}
\alias{get_most_popular_tags}
\alias{get_frequent_tags}
\alias{get_recommended_tags}
\title{Tags}
\usage{
get_most_popular_tags(language = "english")

get_frequent_tags()

get_recommended_tags()
}
\arguments{
\item{language}{Language string of the returned information.}
}
\value{
A dataframe containing all tags sorted by their popularity
}
\description{
Most generally, retrieves a list of tags. \code{get_most_popular_tags}
fetches the most popular tags on Steam at the time. \code{get_frequent_tags}
is an authenticated function that fetches the authenticated user's most
popular tags. \code{get_recommended_tags} returns a list of tags recommended
by the Steam algorithm.
}
\examples{
\dontrun{
get_most_popular_tags(language = "DE")

# requires authentication
auth_credentials("username")
get_frequent_tags()
get_recommended_tags()
}
}
