% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_in_review.R
\name{user_in_review}
\alias{user_in_review}
\alias{friends_in_review}
\alias{achievements_in_review}
\alias{screenshots_in_review}
\title{Year in review}
\usage{
user_in_review(
  steamid,
  year = 2023,
  force_regenerate = FALSE,
  access_source = FALSE
)

friends_in_review(steamid, year = 2023, return_private = FALSE)

achievements_in_review(steamid, appids, year = 2023, total_only = FALSE)

screenshots_in_review(steamid, appids, year = 2023)
}
\arguments{
\item{steamid}{SteamID of a user. The SteamID must be in a format that can
be converted by \code{\link{convert_steamid}}. This includes vanity,
Steam64, Steam2, and Steam3 IDs.}

\item{year}{Review year to return data for. Must be a year smaller than
the current year, otherwise the response will be empty.}

\item{force_regenerate}{Whether to regenerate the year in review or re-use
a previously generated summary.}

\item{access_source}{Unknown.}

\item{return_private}{Unknown.}

\item{appids}{A vector of multiple appIDs of applications in the Steam store.
There are generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query applications programmatically, e.g. using \code{\link{search_apps}}}
}}

\item{total_only}{If \code{TRUE}, returns only a summary list of
achievements in \code{year}.}
}
\value{
\describe{
 \item{\code{user_in_review}}{A nested list containing lots of information
 about playtime, performance, and game acquisitions.}

 \item{\code{friends_in_review}}{A dataframe containing all newly befriended
 Steam friends. The columns\code{privacy_state} and
 \code{rt_privacy_updated} tell about the friends' privacy state.}

 \item{\code{achievements_in_review}}{A dataframe with one row per appID.
 \code{all_time_unlocked_achievements} and \code{unlocked_more_in_future}
 tell about past and future achievements. \code{achievements} is a nested
 dataframe that contains all achievement for a specific game in \code{year}}
}
}
\description{
Get retrospective user-related summaries about past years.
}
\examples{
\dontrun{
steamid <- "76561197960435530"

# get all sorts of summary data from 2023
user_in_review(steamid, year = 2023)

# all new friends in 2023
# authentication necessary
friends_in_review(steamid, year = 2023)

# get a list of all achievements in 2023
achievements_in_review(steamid, appids = c(440, 730), year = 2023)

# return a summary of total achievements in 2023
achievements_in_review(
  steamid,
  appids = c(440, 730),
  year = 2023,
  total_only = TRUE
)

# return a summary of screenshots in 2023
# actually, this returns an empty tibble because I couldn't
# find an example of a working user/app combination
screenshots_in_review(steamid, year = 2023)
}
}
