% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{usrstats}
\alias{usrstats}
\title{User stats for a game}
\usage{
usrstats(key, gameid = "440", steamid = "76561197972495328")
}
\arguments{
\item{key}{Steam API key}

\item{gameid}{Steam game ID}

\item{steamid}{Steam user ID}
}
\description{
Returns list for user stats for a game
}

