% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing.R
\name{wba_active_marketing_messages}
\alias{wba_active_marketing_messages}
\title{Marketing}
\usage{
wba_active_marketing_messages(country = NULL, anonymous_user = FALSE)
}
\arguments{
\item{country}{ISO 3166 country code representing the country from
which to view the Steam store. A list of Steam countries can be retrieved
with \code{\link{get_country_list}}.}

\item{anonymous_user}{If \code{TRUE}, will not return any marketing messages
that require ownership, playtime or wishlisting.}
}
\value{
A dataframe containing the groupID
}
\description{
Marketing messages are sometimes shown to the user as a popup window. This
function can retrieve details about the marketing messages currently active
in a given country.
}
