% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{wba_apps}
\alias{wba_apps}
\title{App list}
\usage{
wba_apps()
}
\value{
A dataframe containing the appID and name of each application in
the Steam store. The app list is cached on disk for a short while until
going stale.
}
\description{
Get a list of applications in the Steam store, either a complete list
or the top games by concurrent players, all-time players, or top releases.
}
\examples{
\donttest{# return all apps on Steam
wba_apps()}
}
\seealso{
\code{\link{steamspy}} for a similar approach by the SteamSpy API

\code{\link{stf_search_apps}} to filter apps using search strings

\code{\link{wba_query}} to filter apps using a filter query
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{wba_apps} \tab no \tab no
}




}
