% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{wba_charts}
\alias{wba_charts}
\alias{wba_charts_weekly}
\alias{wba_charts_ccu}
\alias{wba_charts_total}
\alias{wba_charts_releases}
\title{Charts}
\usage{
wba_charts_weekly(
  country = NULL,
  context = store_context(),
  data_request = store_data_request(),
  start_date = NULL
)

wba_charts_ccu(context = store_context(), data_request = store_data_request())

wba_charts_total()

wba_charts_releases()
}
\arguments{
\item{country}{ISO 3166 country code representing the country from
which to view the Steam store. A list of Steam countries can be retrieved
with \code{\link{get_country_list}}.}

\item{context}{Object of class
\code{\link[=store_context]{StoreBrowseContext}} that specifies the
geographic context from which to access the store. Defaults to a global
Steam realm (\code{steam_realm = 1}), English as the store language
(\code{language = "english"}) and the USA as the access country
(\code{country_code = "US"}).}

\item{data_request}{Object of class
\code{\link[=store_data_request]{StoreBrowseDataRequest}} that specifies
the additional information to be included. Defaults to no extra info.}

\item{start_date}{Date-time object giving the earliest time for which to
retrieve top sellers. Results will be returned for the time between
the start date and the start date plus seven days.}
}
\value{
\itemize{
 \item{\code{wba_charts_weekly}: A dataframe containing the first 100 ranks,
 the appID as well as detailed information about the ranking history. Also
 includes item info as returned by \code{\link{wba_store_items}}. The
 actual start date of the weekly count is included in the \code{start_date}
 attribute.}
 \item{\code{wba_charts_ccu}: A dataframe containing the first 100 ranks,
 their appID, concurrent players and peak players.}
 \item{\code{wba_charts_total}: A dataframe containing the first 100 ranks,
 their appID, rank change, and peak players.}
 \item{\code{wba_charts_releases}: A dataframe containing three rows
 representing the last three months. Each row contains the appIDs of the
 top 88 new releases of that month.}
}
}
\description{
Retrieve Steam (weekly) charts in terms of concurrent players (CCU), total
most played games, and top new releases of the last three months.
}
\details{
Although a store context can be passed to \code{wba_charts_ccu}, it does not
seem to be taken into account. To return a regionalized version of the
Steam charts, you may use \code{wba_charts_weekly}.
}
\examples{
\donttest{# get most played games
wba_charts_total()

# worldwide weekly top sellers
get_weekly_top_sellers()

# weeky top sellers of South Korea
get_weekly_top_sellers("KR")

# top sellers of last month
get_weekly_top_sellers(start_date = Sys.time() - 2629800)}
}
\seealso{
\code{\link{wba_apps}} to return a list of all apps

\code{\link{stf_frontpage}} to
retrieve (genre) charts as shown on the store front page
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{wba_charts_ccu}      \tab no \tab no\cr
    \code{wba_charts_ccu}      \tab no \tab no\cr
    \code{wba_charts_total}    \tab no \tab no\cr
    \code{wba_charts_releases} \tab no \tab no
}




}
