% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcode.R
\name{wba_html_to_bbcode}
\alias{wba_html_to_bbcode}
\title{HTML to BBCode}
\usage{
wba_html_to_bbcode(x, preserve_newlines = FALSE)
}
\arguments{
\item{x}{A string containing HTML formatting.}

\item{preserve_newlines}{If \code{TRUE}, preserves linebreaks. If
\code{FALSE}, also preserves linebreaks -- or so it seems.}
}
\value{
A character string containing BBCode formatted text.
}
\description{
Converts an HTML string to BBCode. BBCode is commonly used in the Steam
forum and other bulletin boards.
}
\examples{
\donttest{string1 <- "<b>test</b>"
string2 <- "<br>\n"
string3 <- "<a href='google.com'>link</a>"

html_to_bbcode(string1)
html_to_bbcode(string2)
html_to_bbcode(string3)}
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{wba_html_to_bbcode} \tab no \tab no
}




}
