% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamid.R
\name{wba_resolve_vanity}
\alias{wba_resolve_vanity}
\title{Get User ID}
\usage{
wba_resolve_vanity(name, type = "profile")
}
\arguments{
\item{name}{Name / Vanity ID of a user account, group or game hub.
See details and examples.}

\item{type}{Type of Steam name. \code{profile} returns the Steam ID of
a user profile, \code{group} returns the ID of a public group and
\code{game_group} returns the ID of a game's official game hub.}
}
\value{
A length-1 character vector containing the Steam ID corresponding
to the input name.
}
\description{
Retrieves the Steam ID64 of a user profile, group or game hub based on
vanity IDs.
}
\details{
There are various way of retrieving vanity IDs depending on the type of
vanity URL type. The vanity URL of a Steam user is the account name
(not the display name). It can be retrieved by inspecting the profile URL:
\preformatted{https://steamcommunity.com/id/{vanity_id}/}

Vanity IDs of groups can be retrieved in a similar way:
\preformatted{https://steamcommunity.com/groups/{vanity_id}/}

Vanity IDs of game hubs are not easily locatable as game hubs are
closely linked to store pages. They can be found by inspecting the
source code of a game page and searching for \code{VANITY_ID}. Vanity IDs
of game hubs are usually the application ID or an abbreviation of the
original title, e.g. \code{dota2} for DOTA 2, \code{TF2} for Team Fortress 2
or simply \code{70} for Half-Life
}
\examples{
\dontrun{
# get user ID
wba_resolve_vanity("gabelogannewell")

# get group ID
wba_resolve_vanity("SteamDB", type = "group")

# get game hub ID
wba_resolve_vanity("TF2", type = "game_group")
}
}
\seealso{
\code{\link{convert_steamid}} for converting from and to vanity.

\code{\link{is_vanity}} to check if a string could be a vanity ID.
}
\section{Authentication}{
The functions of this reference page are subject to the following authentication requirements (Key = API key needed, Login = user login needed):
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{wba_resolve_vanity} \tab yes \tab no
}

API keys can be set by storing the API key in the \code{STEAM_API_KEY} environment variable.


}
