% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{wba_store_items}
\alias{wba_store_items}
\alias{wba_hardware_items}
\title{Get store items}
\usage{
wba_store_items(items, context = NULL, data_request = NULL)

wba_hardware_items(items, context = NULL)
}
\arguments{
\item{items}{A dataframe, list, or character vector containing item IDs. If
a vector, interprets it as appIDs. If a list or dataframe is passed, each
column or list element can refer to a different type of item ID. Available
types are \code{appid}, \code{packageid}, \code{bundleid}, \code{tagid},
\code{creatorid}, and \code{hubcategoryid}. In this way, multiple items
with multiple ID types can be specified. For \code{get_hardware_items},
only vectors are allowed and they are always interpreted as packageIDs.}

\item{context}{Object of class
\code{\link[=store_context]{StoreBrowseContext}} that specifies the
geographic context from which to access the store. Defaults to a global
Steam realm (\code{steam_realm = 1}), English as the store language
(\code{language = "english"}) and the USA as the access country
(\code{country_code = "US"}).}

\item{data_request}{Object of class
\code{\link[=store_data_request]{StoreBrowseDataRequest}} that specifies
the additional information to be included. Defaults to basic information.}
}
\value{
A dataframe containing the requested information on the input
IDs.
}
\description{
Get information about items in the store. \code{get_hardware_items} more
specifically requests special information about items that require physical
shipping.
}
\examples{
\donttest{# if just an ID is passed, it is assumed to be an appID
wba_store_items(10)

# by passing a dataframe, you can pass multiple types of ID at a time
ids <- data.frame(
  appid = c(10, NA, NA),
  creatorid = c(NA, 10, NA),
  packageid = c(NA, NA, 354231)
)

# request basic item info
wba_store_items(ids)

# request info in german language
wba_store_items(ids, context = store_context(language = "german"))

# request info for the swedish store
wba_store_items(ids, context = store_context(country_code = "SE"))

# request info on operating systems
wba_store_items(ids, data_request = store_data_request("platforms"))

# request special info on hardware
wba_hardware_items(c(354231, 1628580))}
}
\section{Authentication}{
The functions of this reference page do not need any kind of authentication to be used (Key = API key needed, Login = user login needed).
\tabular{lll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login}\cr
    \code{wba_store_items}    \tab no \tab no\cr
    \code{wba_hardware_items} \tab no \tab no
}




}
