% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{appdetails}
\alias{appdetails}
\alias{appdetails_steamcmd}
\alias{appdetails_steamspy}
\title{Get app details}
\usage{
appdetails(
  appids,
  client = NULL,
  filters = NULL,
  country_code = "US",
  language = "english"
)

appdetails_steamcmd(appid)

appdetails_steamspy(appid)
}
\arguments{
\item{appids, appid}{One or up to 100 application IDs. If more than one app
IDs is passed, the \code{filters} argument must only contain the
\code{price_overview} filter.}

\item{client}{Unknown parameter.}

\item{filters}{Fields to include in the output. Can be any field name that
is returned by this function. If \code{basic}, includes a number of basic
fields.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{language}{ISO639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}
}
\value{
A nested list containing detailed information on a single app.
If multiple app IDs are passed and \code{filters = "price_overview"},
contains price information on up to 100 applications.
}
\description{
Retrieve detailed information on an application in the steam store.

\code{appdetails} uses Steam's unofficial API.

\code{appdetails_steamcmd}
uses the unofficial \href{https://www.steamcmd.net/}{steamcmd API}, which
provides data from the Steam client network (powered by the
\href{https://steam.readthedocs.io/en/stable/api/steam.client.html}{\code{steam}}
Python library).

\code{appdetails_steamspy} uses the unofficial
\href{https://steamspy.com/}{SteamSpy API}, which provides
custom estimates on owners, tags, genres and price dynamics. To get a full
app list from SteamSpy, see \code{\link{steamspy}}.
}
\note{
\code{appdetails} is rate-limited at 200 requests per 5 minutes.

\code{appdetails_steamcmd} is not rate-limited.

\code{appdetails_steamspy} is rate-limited at 1 request per second.
}
\examples{
\dontrun{
appdetails(10, filters = c("recommendations", "metacritic"))
}
}
