% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{appdetails}
\alias{appdetails}
\alias{appdetails_all}
\title{Get app details}
\usage{
appdetails(
  appids,
  client = NULL,
  filters = NULL,
  country_code = "US",
  language = "english"
)

appdetails_all(
  appids,
  client = NULL,
  filters = NULL,
  country_code = "US",
  language = "english",
  progress = TRUE
)
}
\arguments{
\item{appids}{One or up to 100 application IDs. If more than one app IDs is
passed, the \code{filters} argument must only contain the \code{price_overview}
filter.}

\item{client}{Unknown parameter.}

\item{filters}{Fields to include in the output. Can be any field name that
is returned by this function. If \code{basic}, includes a number of basic
fields.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{language}{Language string of the returned information.}
}
\value{
A nested list containing detailed information on a single app.
If multiple app IDs are passed and \code{filters = "price_overview"},
contains price information on up to 100 applications.

\code{appdetails_all} returns a list where each index corresponds to the
app details of an app ID in \code{appids}.
}
\description{
Retrieve detailed information on an application in the steam store. Uses
Steam's unofficial API.

\code{appdetails} returns details on a single app or price information
on up to 100 apps. \code{appdetails} is a simple wrapper that iterates
through \code{appids} and returns a list of details on all apps.
}
\note{
This endpoint is rate-limited at 200 requests per 5 minutes.
}
\examples{
\dontrun{
appdetails(10, filters = c("recommendations", "metacritic"))
}
}
