% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\alias{auth_credentials}
\alias{auth_qr}
\alias{auth_logout}
\alias{auth_history}
\title{Session authentication}
\usage{
auth_credentials(
  username,
  password = openssl::askpass,
  shared_secret = NULL,
  persistent = getOption("steamr_persistent", FALSE),
  friendly_name = "steamr",
  details = "Login using credentials"
)

auth_qr(
  friendly_name = "steamr",
  device_details = "Login using QR code",
  persistent = getOption("steamr_persistent", FALSE)
)

auth_logout()

auth_history()
}
\arguments{
\item{username}{Account name of the user to be authenticated.}

\item{password}{A function that safely prompts the Steam password. Defaults
to \code{\link[askpass]{askpass}} which obfuscates the input. The password
is RSA encrypted before transmission to the Steam servers.}

\item{shared_secret}{A 5-character string representing the one-time Steam
Guard code needed to confirm the login. Note that the Steam Guard code
is regenerated every 30 seconds. If \code{shared_secret} is \code{NULL},
authenticates using two-factor authentication. Two-factor confirmation
requires the manual confirmation of the login request in the Steam Guard
mobile app.}

\item{persistent}{Whether to start a persistent or ephemeral session.
If the session is ephemeral (default), then it is reset when the R session
ends. If it is persistent, then the necessary cookies are stored in a
persistent cache in the file system and recovered when the package is loaded.
\code{auth_credentials} additionally requests a persistent session from the
Steam API indicating that the requested access token is persistent.
\code{auth_credentials} is thus potentially more persistent than
\code{auth_qr}.}

\item{friendly_name}{Name of the session. Used to track authorized devices
in Steam Guard.}

\item{details}{Arbitrary details about the device attempting to
authenticate. Will be shown in Steam Guard.}
}
\value{
An object of class \code{steam_auth_session} holding information
about vanity ID, Steam64 ID, client ID, request ID, and access token. The
object is also attached to the session. \code{auth_logout} returns
\code{NULL}, invisibly. \code{auth_history} returns a dataframe where each
row is a login attempt.
}
\description{
Initializes an authorized Steam session. Some functions in both the
Web API and storefront API require the session to be authenticated. This
concerns all API methods that access personal account information.
\code{auth_*} functions essentially sign in a user programmatically.

\code{auth_qr} authenticates by showing a QR code that can be
scanned using the Steam mobile app. The QR code is shown as an R plot
and refreshes every 5 seconds.

\code{auth_logout} ends the active authenticated session by formally logging
out of Steam and removing all cookies and authentication information from
the session and the cache.
}
\details{
Internally, \code{auth_credentials} requests a public RSA key using the
\code{GetPasswordRSAPublicKey} method from the \code{IAuthenticationService}
interface. The public key is used to encrypt the entered password to safely
transfer it to Steam. An authenticated session is requested by querying
\code{BeginAuthSessionViaCredentials}. If a captcha is required, the
function aborts. If no Steam Guard code is provided, the function will
attempt to authenticate using 2 factor authentication. After confirming
the authentication, \code{auth_credentials} will redirect to all Steam
websites requiring authentication to set the necessary cookies.

\code{auth_qr} follows a very similar process, but requests an authenticated
session using \code{BeginAuthSessionViaQR}. The resulting challenge URL
is converted to a QR code using \code{\link[qrcode]{qrcode}}. Scanning the
QR code with the Steam Guard mobile app automatically authenticates the
session.
}
\note{
Session authentication is only possible in interactive mode because it
requires the manual insertion of Steam guard codes, confirmation of
login requests in the mobile app, or the scanning of QR codes. All of these
actions are not suitable for batch processing.
}
\examples{
\dontrun{
user <- "gabelogannewell"
auth_credentials(user)

# use a different password method
auth_credentials(user, password = rstudioapi::askForPassword)

# set a friendly name to identify this session
auth_credentials(user, friendly_name = "R session")

# skip two-factor authentication by providing a Steam Guard code
auth_credentials(user, shared_secret = "XXXXX")

# authenticating a persistent session survives restarts of R
auth_credentials(user, persistent = TRUE)
.rs.restartR() # restart R
is_authenticated() # returns TRUE

# sign in using a QR code
auth_qr()

# the new login attempts can be retrieved using `auth_history()`
auth_history()

# end the authenticated session by revoking the refresh token
auth_logout()}
}
\section{Authentication}{
The functions of this reference page are subject to the following authentication requirements (Key = API key needed, Login = user login needed):
\tabular{llll}{
    \strong{Function} \tab \strong{Key} \tab \strong{Login} \tab \strong{Note}\cr
    \code{auth_credentials} \tab no \tab no  \tab                  \cr
    \code{auth_qr}          \tab no \tab no  \tab                  \cr
    \code{auth_logout}      \tab no \tab yes \tab Ends login status\cr
    \code{auth_history}     \tab no \tab yes \tab                  
}



To learn more about user authentication, see \code{\link{auth_credentials}}.
}
