% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\title{Common arguments}
\arguments{
\item{appid}{Application ID referencing a Steam application. There are
generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query appIDs programmatically, e.g. using \code{\link{search_apps}}}
}}

\item{appids}{A vector of multiple appIDs of applications in the Steam store.
There are generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query applications programmatically, e.g. using \code{\link{search_apps}}}
}}

\item{steamid}{SteamID of a user. The SteamID must be in a format that can
be converted by \code{\link{convert_steamid}}. This includes vanity,
Steam64, Steam2, and Steam3 IDs.}

\item{steamids}{A vector of multiple SteamIDs of a user. The SteamIDs must
be in a format that can be converted by \code{\link{convert_steamid}}.
This includes vanity, Steam64, Steam2, and Steam3 IDs.}

\item{paginate}{If \code{TRUE}, paginates through the results. Some methods
provide data access in digestible chunks and do not return all available
information at once. The \code{paginate} automatically retrieves all pages
until \code{max_pages} is reached.}

\item{max_pages}{Maximum number of pages to paginate. Ignored if
\code{paginate} is \code{FALSE}. Defaults to \code{Inf} such that all
available data are paginated.}
}
\description{
Common arguments that are reused in various functions of the Web and
storefront API.
}
\keyword{internal}
