% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{common}
\alias{common}
\title{Common arguments}
\arguments{
\item{appid}{A single appID of an application in the Steam store.}

\item{appids}{A vector of multiple appIDs of applications in the Steam store.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{elanguage}{Numeric code representing the store language. A list
of language codes and their corresponding languages is defined in
\code{\link{ELanguage}}.}

\item{country_code}{ISO 3166 country code representing the country from
which to view the Steam store.}

\item{steam_realm}{Number describing the Steam realm. A value of 1
indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.}

\item{steamid}{SteamID of a user. The SteamID must be in a format that can
be converted by \code{\link{convert_steamid}}. This includes vanity,
Steam64, Steam2, and Steam3 IDs.}

\item{steamids}{A vector of multiple SteamIDs of a user. The SteamIDs must
be in a format that can be converted by \code{\link{convert_steamid}}.
This includes vanity, Steam64, Steam2, and Steam3 IDs.}

\item{paginate}{If \code{TRUE}, paginates through the results. Some methods
provide data access in digestible chunks and do not return all available
information at once. The \code{paginate} automatically retrieves all pages
until \code{max_pages} is reached.}

\item{max_pages}{Maximum number of pages to paginate. Ignored if
\code{paginate} is \code{FALSE}. Defaults to \code{Inf} such that all
available data are paginated.}
}
\description{
Common arguments that are reused in various functions of the Web and
storefront API.
}
\keyword{datasets}
