% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\docType{data}
\name{games}
\alias{games}
\title{Steam Game Data}
\format{
A data frame with 500 rows and 11 variables:
\describe{
\item{app_id}{Unique Steam ID}
\item{name}{Game name}
\item{description}{Short game description}
\item{image}{Image URL}
\item{developers, publishers}{Game developers and publishers}
\item{price}{Price in EUR}
\item{metacritic}{Metacritic score}
\item{genres}{Game genres (comma-separated)}
\item{recommendations}{Number of recommendations}
\item{release_date}{Game release date}
}
}
\source{
Huggingface: \href{https://huggingface.co/datasets/FronkonGames/steam-games-dataset}{FronkonGames/steam-games-dataset}

Steam API: \url{https://store.steampowered.com/api/appdetails?appids=AppID}
}
\usage{
games
}
\description{
Subset of game details from the Steam API and Huggingface dataset.
}
\keyword{datasets}
