% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_app_list}
\alias{get_app_list}
\title{Get the complete app list}
\usage{
get_app_list()
}
\value{
A list containing the name and id of every program
  available in the Steam store.
}
\description{
Gets the name and id of every program available in the
  Steam store.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{appid}{The app's id}
  \item{name}{The app's name}
}
}

