% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_app_list}
\alias{get_app_list}
\title{Get complete app list}
\usage{
get_app_list()
}
\value{
the complete app list available in the steam
  store
}
\description{
Pull the complete app list. Each item contains the
  title and app ID of each program available in the
  steam store.
}

