% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_app_list}
\alias{get_app_list}
\title{Get complete app list}
\usage{
get_app_list()
}
\value{
The complete app list available in the Steam
  store.

Each item in the resulting list contains the app's AppID
   and the app's name.
}
\description{
Pull the complete app list. Each item contains the
  title and app ID of each program available in the
  Steam store.
}

