% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-app-list.R
\name{get_app_list}
\alias{get_app_list}
\title{Get the complete app list}
\usage{
get_app_list()
}
\value{
A list containing the name and id of every program
  available in the Steam store.
}
\description{
Gets the name and id of every program available in the
  Steam store.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{appid}{The app's AppID}
  \item{name}{The app's name}
}
}

