% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app_list}
\alias{get_app_list}
\alias{get_games_by_ccu}
\alias{get_most_played_games}
\alias{get_top_releases}
\alias{get_apps_in_genre}
\alias{get_apps_in_category}
\title{App list}
\usage{
get_app_list()

get_games_by_ccu(
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE
)

get_most_played_games()

get_top_releases()

get_apps_in_genre(genre, language = "english", country_code = "US")

get_apps_in_category(category, language = "english", country_code = "US")
}
\value{
\describe{
 \item{\code{get_app_list}}{A dataframe containing the appID and name.}

 \item{\code{get_games_by_ccu}}{A dataframe containing the appID, the
 rank by CCU and the total number and all-time record of concurrent players.}

 \item{\code{get_most_played_game}}{A dataframe containing the appID, the
 rank by total all-time players and players last week as well as the
 all-time record of concurrent players.}

 \item{get_top_releases}{A dataframe with three rows containing the top
 releases of the last three months. Each row contains the start of the
 month, the url path for the top releases and a vector top release appIDs.
 The URL path can be appended to the following URL:

 \code{https://store.steampowered.com/charts/topnewreleases/}}
}
}
\description{
Get a list of applications in the Steam store, either a complete list
or the top games by concurrent players, all-time players, or top releases.
}
\examples{
get_app_list()

# get most played games
get_games_by_ccu()

# get most popular games of all time
get_most_played_games()

# get the best releases
get_top_releases()
}
\seealso{
\code{\link{steamspy}} for a similar approach by the SteamSpy API
}
