% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seurat.R
\name{get_assay_data}
\alias{get_assay_data}
\title{Internal helper for Seurat v4/v5 data access}
\usage{
get_assay_data(
  se,
  assay = SeuratObject::DefaultAssay(se),
  what = c("counts", "data", "scale.data")
)
}
\description{
Uses \code{layer} for SeuratObject >= 5, else \code{slot}.
}
\keyword{internal}
