% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_auth}
\alias{get_auth}
\title{Get authentication}
\usage{
get_auth(field = NULL)
}
\value{
An auth object as returned by \code{auth_credentials} and
\code{auth_qr}.
}
\description{
Retrieve global authentication object. When authenticating a session
using \code{\link{auth_credentials}} or \code{\link{auth_qr}}, an auth
object is attached to the R session. This object can be useful when
manually creating requests to the Steam API. Use this function to retrieve
these informations. Only for advanced usage.
}
\examples{
\dontrun{
auth_credentials("name")

get_auth()
}
}
