% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_friend_list}
\alias{get_friend_list}
\title{Get User's friend list}
\usage{
get_friend_list(steam_key, steam64_id, relationship)
}
\arguments{
\item{steam_key}{a Steam Web API Key.}

\item{steam64_id}{A 64-bit SteamID to return friend list for.}

\item{relationship}{Relationship filter. Possibles values: \code{all}, \code{friend}.}
}
\value{
A user's friend list
}
\description{
Returns the friend list of a user, provided his or her
  Steam Community profile visibility is set to "Public".
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{steamid}{The friend's 64-bit SteamID}
  \item{relationship}{The relationship qualifier}
  \item{friend_since}{UNIX timestamp representing the time
     when the relationship was created}
}
}

\section{Steam Web API Key}{

Get a Steam Web API Key at \url{http://steamcommunity.com/dev/apikey}.
}

