% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievements.R
\name{get_game_achievements}
\alias{get_game_achievements}
\title{Achievements}
\usage{
get_game_achievements(appid, language = "english")
}
\arguments{
\item{appid}{appID of an application in the Steam store.}

\item{language}{ISO639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}
}
\value{
A dataframe containing information about a game's achievements.
}
\description{
Retrieve achievements for a game.
}
