% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_global_achievement_percentages_for_app}
\alias{get_global_achievement_percentages_for_app}
\title{Get Global Achievement Percentages for App}
\usage{
get_global_achievement_percentages_for_app(app_id)
}
\arguments{
\item{app_id}{The app's AppID.}
}
\value{
The latest global achievement percentages for an app.

Each item in the resulting list contains the achievement name
   and the percentage.
}
\description{
Get the global achievement percentages for an app.
}
\examples{
# get global achievement percentages for Super Meat Boy
get_global_achievement_percentages_for_app(40800)
}

