% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_browse.R
\name{get_items}
\alias{get_items}
\title{Get store items}
\usage{
get_items(
  items,
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE
)
}
\arguments{
\item{items}{Store item or list of store items as returned by
\code{\link{store_item}}.}

\item{language}{Language string of the returned information.}

\item{elanguage}{Language identifier of the returned information.}

\item{country_code}{ISO-2 country code as returned by
\code{\link{get_countries}}. The Steam store differs depending on the
country it is accessed from.}

\item{steam_realm}{Number describing the Steam realm.}

\item{include}{List of extra information to include. Can be one or several
of the following: \code{release}, \code{platforms},
\code{all_purchase_options}, \code{screenshots}, \code{trailers},
\code{ratings}, \code{tag_count}, \code{reviews}, \code{basic_info},
\code{supported_languages}, \code{full_description}, \code{included_items}
and \code{assets_without_overrides}. If \code{all}, returns all information.}

\item{apply_user_filters}{t}
}
\value{
A dataframe containing the requested information on the input
IDs.
}
\description{
Get information about items in the store
}
\examples{
\dontrun{
ids <- list(
  store_item(268770),
  store_item(1182620),
  store_item(creatorid = 10)
)

# request basic item info
get_items(ids)

# request info in german language
get_items(ids, language = "german")

# request info for the swedish store
get_items(ids, country_code = "SE")

# request info on operating systems
get_items(ids, include = "platforms")

# request all info
get_items(ids, include = "all")
}

}
\seealso{
Other store: 
\code{\link{get_hardware_items}()},
\code{\link{get_store_categories}()},
\code{\link{store_item}()}
}
\concept{store}
