% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{get_news}
\alias{get_news}
\title{News}
\usage{
get_news(
  appid,
  maxlength = NULL,
  enddate = NULL,
  count = 20,
  feeds = NULL,
  tags = NULL
)
}
\arguments{
\item{appid}{Application ID referencing a Steam application. There are
generally two ways of retrieving an appID:
\itemize{
 \item{Inspect or parse the store URL of an application, e.g. \code{https://store.steampowered.com/app/10/CounterStrike/}}
 \item{Query appIDs programmatically, e.g. using \code{\link{search_apps}}}
}}
}
\description{
Get news data for a given application.
}
