% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_news_for_app}
\alias{get_news_for_app}
\title{Get news for an app}
\usage{
get_news_for_app(app_id, count, max_length)
}
\arguments{
\item{app_id}{The app's AppID.}

\item{count}{The number of news entries to return.
Returns all news entries if unspecified.}

\item{max_length}{The maximum length of each news entry.
  Returns all characters if unspecified.}
}
\value{
The latest news of a game specified by its AppID.

Each item in the resulting list contains: a gid, a title, a url,
a shortened excerpt of the contents (to \code{max_length} characters),
terminated by "..." if longer than \code{max_length}, a feed label
a UNIX timestamp date, an author, and the feedname.
}
\description{
Get the latest news for an app.
}
\examples{
# get the most recent The Binding of Isaac: Rebirth news post
get_news_for_app(250900, count = 1)

# get all SteamWorld Dig news posts
get_news_for_app(252410)
}

