% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_profile_items}
\alias{get_profile_items}
\title{Profiles}
\usage{
get_profile_items(
  steamid,
  language = "english",
  status = c("owned", "equipped")
)
}
\arguments{
\item{steamid}{Steam ID of a user for which profile information is to be
fetched. The Steam ID must be in a format that can be converted by
\code{\link{convert_steamid}}.}

\item{language}{If text data is returned, specify the text language.}
}
\description{
Fetch cosmetic information from a user profile.
}
