% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_steam_id_64}
\alias{get_steam_id_64}
\title{Get a 64-bit Steam ID}
\usage{
get_steam_id_64(steam_key, steam_id)
}
\arguments{
\item{steam_key}{a Steam Web API Key.}

\item{steam_id}{a Steam ID}
}
\value{
a 64-bit Steam ID
}
\description{
Get a 64-bit Steam ID
}
\section{Steam Web API Key}{

Get a Steam Web API Key at \url{http://steamcommunity.com/dev/apikey}
}

