% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_browse.R
\name{get_store_categories}
\alias{get_store_categories}
\title{Get store categories}
\usage{
get_store_categories(language = "english", elanguage = NULL)
}
\arguments{
\item{language}{Language string of the returned information.}

\item{elanguage}{Language identifier of the returned information.}
}
\value{
A dataframe containing names and information on all
store categories.
}
\description{
Get a list of Steam store categories.
}
\examples{
\dontrun{
get_store_categories(language = "swedish")
}
}
\seealso{
Other store: 
\code{\link{get_hardware_items}()},
\code{\link{get_items}()},
\code{\link{store_item}()}
}
\concept{store}
