% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_user_stats_for_game}
\alias{get_user_stats_for_game}
\title{Get User Stats For Game.}
\usage{
get_user_stats_for_game(steam_key, steam64_id, app_id)
}
\arguments{
\item{steam_key}{A Steam Web API Key.}

\item{steam64_id}{A 64-bit SteamID to return friend list for.}

\item{app_id}{The app's AppID.}
}
\value{
A list of achievements for a user by AppID.
}
\description{
Returns a list of achievements for a user by AppID.
}
\section{Returns}{

Each item in the resulting list contains:
\describe{
  \item{apiname}{The API name of the achievement}
  \item{achieved}{Whether or not the achievement has been completed}
}
}

\section{Steam Web API Key}{

Get a Steam Web API Key at \url{http://steamcommunity.com/dev/apikey}
}
\seealso{
\code{\link{get_player_achievements}}
}

