% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_weekly_top_sellers}
\alias{get_weekly_top_sellers}
\title{Weekly top sellers}
\usage{
get_weekly_top_sellers(
  country,
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE,
  start_date = NULL,
  page_count = NULL
)
}
\arguments{
\item{country}{ISO-2 code of the country for which to return top sellers.}

\item{language}{ISO 639-1 language code all tokenized strings should be
returned in. Formally, a tokenized string is a string that is prefixed
with a \code{#} in the VDF file. Not all tokenized strings have a translation
for all languages. If no translation is available, defaults to English.}

\item{elanguage}{Numeric code representing the store language. A list
of language codes and their corresponding languages is defined in
\code{\link{ELanguage}}.}

\item{country_code}{ISO 3166 country code representing the country from
which to view the Steam store.}

\item{steam_realm}{Number describing the Steam realm. A value of 1
indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.}

\item{start_date}{Date-time object giving the earliest time for which to
retrieve top sellers. Results will be returned for the time between
the start date and the start date plus seven days.}

\item{page_count}{Maximum pages to return. Defaults to one page.}

\item{page_start}{On which part to start looking. Defaults to page 1.}
}
\value{
A dataframe containing information on weekly top sellers.
Metadata is stored in attributes: \code{timestamp} holds the earliest
time of the week, \code{next_page_start} specifies where the next page
would start.
}
\description{
Get a list of weekly top sellers by country.
}
\note{
This function currently does not work as intended. The
\code{GetWeekyTopSellers} method does not respond to any parameters. The
only use of this function at the moment is to retrieve the top 20 world-wide
bestsellers at the time.
}
\seealso{
\code{\link{get_country_list}}
}
