% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_permutation_matrices.R
\name{load_permutation_matrices}
\alias{load_permutation_matrices}
\title{Load Permutation Matrices}
\usage{
load_permutation_matrices(file_paths)
}
\arguments{
\item{file_paths}{A list of file paths to the permutation matrices.}
}
\value{
A named list of loaded permutation matrices.
}
\description{
This function loads permutation matrices from given file paths.
}
