% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metaclusters.R
\name{plot_metaclusters}
\alias{plot_metaclusters}
\title{Plot Cluster Distributions for MetaCells or MetaSpots}
\usage{
plot_metaclusters(metaspots)
}
\arguments{
\item{metaspots}{A list containing:
\describe{
\item{cell_types}{Character vector of cell type names.}
\item{metacells}{List of Seurat objects corresponding to each cell type.}
}}
}
\value{
Invisible NULL. The function is called for its side effect of plotting
and printing cluster summaries.
}
\description{
This function takes a list containing \code{cell_types} and \code{metacells} (Seurat objects)
and plots histograms of the number of units (cells or spots) per cluster
for each type. It also prints summary information to the console.
}
\examples{
\dontrun{
plot_metaclusters(sorted_metaspots)
}

}
