% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{query_locations}
\alias{query_locations}
\title{Country info}
\usage{
query_locations(country = NULL, region = NULL)
}
\arguments{
\item{country}{ISO-2 country code. If specified, returns information
on regions, otherwise countries.}

\item{region}{Region code within a country specified in \code{country}.
Country must not be \code{NULL} if \code{region} is specified. If specified,
returns information on cities, otherwise on regions.}
}
\value{
A dataframe with information on country code, region code, city
code and geographical name.
}
\description{
Retrieves information on Steam's country, region, and city codes.
}
\examples{
\dontrun{
# get country codes
query_locations()

# get region codes of Germany
query_locations(country = "DE")

# get city codes of Baden-Wurttemberg, Germany
query_locations(country = "DE", region = "01")
}
}
