% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{query_locations}
\alias{query_locations}
\alias{get_country_list}
\title{Location}
\usage{
query_locations(country = NULL, region = NULL)

get_country_list(language = "english")
}
\arguments{
\item{country}{ISO-2 country code. If specified, returns information
on regions, otherwise countries.}

\item{region}{Region code within a country specified in \code{country}.
Country must not be \code{NULL} if \code{region} is specified. If specified,
returns information on cities, otherwise on regions.}
}
\value{
\describe{
 \item{\code{query_locations}}{A dataframe with country code, name,
 and whether the location can be further subdivided into states or cities.}

 \item{\code{get_country_list}}{A dataframe containing country code
 and country name.}
}
}
\description{
Retrieves information on Steam's country, region, and city codes.

\code{query_locations} queries the storefront API and can drill down all
Steam location levels, while \code{get_country_list} can only provide
information on the country level.
}
\examples{
\dontrun{
# get country codes
query_locations()

# get region codes of Germany
query_locations(country = "DE")

# get city codes of Baden-Wurttemberg, Germany
query_locations(country = "DE", region = "01")

# get localized country names
get_country_list(language = "german")
}
}
