% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-definitions-args.R
\docType{data}
\name{steam_api_args}
\alias{ARG_APP_ID}
\alias{ARG_APP_ID_PLAYING}
\alias{ARG_COUNT}
\alias{ARG_FORMAT}
\alias{ARG_GAME_ID}
\alias{ARG_INCLUDE_APP_INFO}
\alias{ARG_INCLUDE_FREE_GAMES_INFO}
\alias{ARG_KEY}
\alias{ARG_MAX_LENGTH}
\alias{ARG_RELATIONSHIP}
\alias{ARG_STEAM_ID}
\alias{ARG_STEAM_IDS}
\alias{ARG_VANITY_URL}
\alias{steam_api_args}
\title{Steam Web API service definition argument constants}
\format{a character string}
\source{
\url{https://developer.valvesoftware.com/wiki/Steam_Web_API}
}
\usage{
ARG_APP_ID

ARG_GAME_ID

ARG_MAX_LENGTH

ARG_COUNT

ARG_FORMAT

ARG_KEY

ARG_STEAM_IDS

ARG_STEAM_ID

ARG_RELATIONSHIP

ARG_VANITY_URL

ARG_INCLUDE_APP_INFO

ARG_INCLUDE_FREE_GAMES_INFO

ARG_COUNT

ARG_APP_ID_PLAYING
}
\description{
Service definition argument constants for Steam Web API
}
\seealso{
\code{\link{steam_api_service_definitions}} for service definition
   constants.
}
\keyword{datasets}

