% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{steam_api_service_definitions}
\alias{STEAM_API_APP_LIST}
\alias{STEAM_API_GLOBAL_ACHIEVEMENT}
\alias{STEAM_API_NEWS}
\alias{steam_api_service_definitions}
\title{Steam Web API service definition constants}
\format{\preformatted{ chr "http://api.steampowered.com/ISteamApps/GetAppList/v2/"
}}
\usage{
STEAM_API_APP_LIST

STEAM_API_NEWS

STEAM_API_GLOBAL_ACHIEVEMENT
}
\description{
Service definition constants for Steam Web API
}
\keyword{datasets}

