% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{steam_stats}
\alias{steam_stats}
\alias{get_supported_api_list}
\alias{get_servertime}
\alias{get_servers}
\title{Meta info}
\usage{
steam_stats()

get_supported_api_list()

get_servertime()

get_servers(filter = NULL, limit = NULL)
}
\value{
\describe{
 \item{\code{get_supported_api_list()}}{An unnamed list where each
 index contains a named list with information on the respective API
 endpoint.}

 \item{\code{get_servertime()}}{A \code{POSIXct} value.}

 \item{\code{steam_stats()}}{A named list of length 2}

 \item{\code{get_servers()}}{A dataframe containing information about
 available game servers.}

}
}
\description{
Fetch metadata about the Steam server.

\code{get_supported_api_list()} returns information about all supported
API endpoints of the web API. A complete list including undocumented
endpoints can be found on \url{https://steamapi.xpaw.me/}.

\code{get_servertime()} returns the current Steam server time.

\code{steam_stats()} returns the current users online and ingame.
}
\examples{
\dontrun{
get_supported_api_list()

get_servertime()

steam_stats()

get_servers()
}
}
