% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamr-package.R
\docType{package}
\name{steamr-package}
\alias{steamr}
\alias{steamr-package}
\title{steamr: Visualising and processing video game data from steam}
\description{
TODO expand with function information: This package can take data from the steam api (as a dataframe) and offers several ways of processing this data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxikuehn/steamr}
  \item \url{https://maxikuehn.github.io/steamr/}
  \item Report bugs at \url{https://github.com/maxikuehn/steamr/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Kühn \email{3009677@stud.hs-mannheim.de}

Authors:
\itemize{
  \item Cedric Mingham \email{2021650@stud.hs-mannheim.de}
  \item Katharina Salewski \email{1911459@stud.hs-mannheim.de}
}

}
\keyword{internal}
