% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamr-package.R
\docType{package}
\name{steamr-package}
\alias{steamr}
\alias{steamr-package}
\title{steamr: Steam Game Data Analysis}
\description{
This package provides functions for analyzing and visualizing Steam game data. It allows users to explore various aspects of Steam games, including prices, genres, and other details. The package includes functionality to manipulate and clean game data and visualize trends in pricing across different genres.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxikuehn/steamr}
  \item \url{https://maxikuehn.github.io/steamr/}
  \item Report bugs at \url{https://github.com/maxikuehn/steamr/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Kühn \email{3009677@stud.hs-mannheim.de}

Authors:
\itemize{
  \item Cedric Mingham \email{2021650@stud.hs-mannheim.de}
  \item Katharina Salewski \email{1911459@stud.hs-mannheim.de}
}

}
\keyword{internal}
