% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/steamR-package.R
\docType{package}
\name{steamR}
\alias{steamR}
\alias{steamR-package}
\title{steamR}
\description{
An interface to Steam's Web API.
}
\details{
This package provides a wrapper for Steam's Web API.
}
\section{Methods}{

See \code{\link{steamR_methods}} for a list of methods available in the steamR package.
}

\section{Steam Web API Key}{

All use of the Steam Web API requires the use of an API Key.

Get a Steam Web API Key at \url{http://steamcommunity.com/dev/apikey}.
}

\section{Steam API Terms of Use}{

Use of Steam's APIs requires that you agree to the \href{http://steamcommunity.com/dev/apiterms}{Steam API Terms of Use}.
}

\section{Steam Web API Documentation}{

Documentation for Steam's Web API can be found \href{https://developer.valvesoftware.com/wiki/Steam_Web_API}{here}.
}
\author{
Jose Gallegos, \email{josegallegos07@gmail.com}
}

