% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{get_tags}
\alias{get_most_popular_tags}
\alias{get_frequent_tags}
\alias{get_recommended_tags}
\title{Tags}
\usage{
get_tags(language = "english", hash = NULL)

get_most_popular_tags(language = "english")

get_frequent_tags()

get_recommended_tags()
}
\arguments{
\item{hash}{Hash included in the \code{hash} attribute of a previous
function call. If \code{hash} matches with the hash of the requested tags,
then a \code{DuplicateRequest} error is thrown.}
}
\value{
All functions return a dataframe containing the tagIDs and tag
names. \code{get_tags} returns all known tags sorted by their tagID.
\code{get_most_popular_tags} returns the most popular tags sorted by
their popularity. \code{get_frequent_tags} returns the most frequent tags
by the authenticated user sorted by the count in the \code{count} column.
\code{get_recommended_tags} returns tags recommended by the Steam algorithm
with a secret sorting.
}
\description{
Retrieves user-provided tags in the Steam store with different emphases
and orders. \code{get_tags} and \code{get_most_popular_tags} operate on tags
in the entire Steam store. \code{get_frequent_tags} and
\code{get_recommended_tags} are specifically based on the authenticated
user.
}
\examples{
\dontrun{
# get all tags
get_tags()

# get the most popular tags translated to German
get_most_popular_tags(language = "DE")

# requires authentication
auth_credentials("username")
get_frequent_tags()
get_recommended_tags()
}
}
